### 2024-06-17 #150 @luraess
@kaipartmann Thanks for submitting your draft to the JuliaCon Proceedings! I will be the managing editor of your paper. Could you please suggest me a succinct list of relevant reviewers and while we are working on getting them on for revision, please take some time fixing the reference issues listed here https://github.com/JuliaCon/proceedings-review/issues/150#issuecomment-2165819842 . To trigger PDF generation, use `@editorialbot generate pdf` command in here.
