### 2024-06-17 4be6677
Update tutorial on static tension
