### 2024-06-17 5a607b4
Minor refactoring
