+++ src/time_solvers/dynamic_relaxation.jl
@@ -43,7 +43,7 @@ function solve!(dh::AbstractDataHandler, dr::DynamicRelaxation,
                 options::AbstractOptions)
     export_reference_results(dh, options)
     Δt = dr.Δt
-    init_mass!(dh, dr)
+    init_density_matrix!(dh, dr)
     if mpi_isroot()
         p = Progress(dr.n_steps; dt=1, desc="TIME INTEGRATION LOOP", color=:normal,
                      barlen=40, enabled=progress_bars())
@@ -56,38 +56,48 @@ function solve!(dh::AbstractDataHandler, dr::DynamicRelaxation,
     return nothing
 end
 
-function init_mass!(dh::AbstractThreadsDataHandler, dr::DynamicRelaxation)
+function init_density_matrix!(dh::AbstractThreadsDataHandler, dr::DynamicRelaxation)
     @threads :static for chunk in dh.chunks
-        _init_mass!(chunk, dr, chunk.paramsetup)
+        _init_density_matrix!(chunk, dr, chunk.paramsetup)
     end
     return nothing
 end
 
-function init_mass!(dh::AbstractMPIDataHandler, dr::DynamicRelaxation)
-    _init_mass!(dh.chunk, dr, dh.chunk.paramsetup)
+function init_density_matrix!(dh::AbstractMPIDataHandler, dr::DynamicRelaxation)
+    _init_density_matrix!(dh.chunk, dr, dh.chunk.paramsetup)
     return nothing
 end
 
-function _init_mass!(chunk::AbstractBodyChunk, dr::DynamicRelaxation,
+function _init_density_matrix!(chunk::AbstractBodyChunk, dr::DynamicRelaxation,
                                params::AbstractPointParameters)
-    mass = chunk.storage.mass
+    system = chunk.system
+    density_matrix = chunk.storage.density_matrix
     for i in each_point_idx(chunk.ch)
-        for j in 1:3
-            mass[j, i] = dr.Λ * 6 * params.K * dr.Δt^2 / (1 / 3 * params.δ^2)
-        end
+        calc_density_matrix!(density_matrix, system, params, dr, i)
     end
     return nothing
 end
 
-function _init_mass!(chunk::AbstractBodyChunk, dr::DynamicRelaxation,
+function _init_density_matrix!(chunk::AbstractBodyChunk, dr::DynamicRelaxation,
                                paramsetup::AbstractParameterHandler)
-    mass = chunk.storage.mass
+    system = chunk.system
+    density_matrix = chunk.storage.density_matrix
     for i in each_point_idx(chunk.ch)
-        params = get_params(chunk, i)
-        for d in 1:3
-            mass[d, i] = dr.Λ * 6 * params.K * dr.Δt^2 / (1 / 3 * params.δ^2)
-        end
+        params = get_params(paramsetup, i)
+        calc_density_matrix!(density_matrix, system, params, dr, i)
     end
+    return nothing
+end
+
+@inline function calc_density_matrix!(density_matrix::Matrix{Float64}, system::BondSystem,
+                                      params::AbstractPointParameters,
+                                      dr::DynamicRelaxation, i::Int)
+    n_bonds = system.n_neighbors[i]
+    k = 5π * params.δ^2 * params.bc
+    Λ = dr.Λ * 1 / 4 * dr.Δt^2 * n_bonds * k
+    density_matrix[1, i] = Λ
+    density_matrix[2, i] = Λ
+    density_matrix[3, i] = Λ
     return nothing
 end
 
[…]
