+++ paper/ref.bib
@@ -70,7 +70,6 @@
 
 @article{Jha2021,
   doi       = {10.21105/joss.03020},
-  url       = {https://doi.org/10.21105/joss.03020},
   year      = {2021},
   publisher = {The Open Journal},
   volume    = {6},
@@ -88,7 +87,7 @@
   number  = {3},
   pages   = {194-204},
   year    = {2010},
-  doi     = {https://doi.org/10.1016/j.tafmec.2010.08.001},
+  doi     = {10.1016/j.tafmec.2010.08.001},
   author  = {Bahattin Kilic and Erdogan Madenci}
 }
 
@@ -156,7 +155,7 @@
 
 @article{Silling2000,
   author   = {S. A. Silling},
-  doi      = {https://doi.org/10.1016/S0022-5096(99)00029-0},
+  doi      = {10.1016/S0022-5096(99)00029-0},
   issn     = {0022-5096},
   journal  = {Journal of the Mechanics and Physics of Solids},
   keywords = {Dynamics, Elastic material, Fracture, Stability and bifurcation, Stress waves},
@@ -210,7 +209,7 @@
   volume   = {21},
   pages    = {101322},
   year     = {2023},
-  doi      = {https://doi.org/10.1016/j.softx.2023.101322},
+  doi      = {10.1016/j.softx.2023.101322},
   author   = {Christian Willberg and Jan-Timo Hesse and Marc Garbade and Martin Rädel and Falk Heinecke and Andreas Schuster and Anna Pernatii},
   keywords = {Peridynamics, User material, UMAT, Interface, Material modeling, Damage model}
 }
