### 2024-06-17 #150 @kaipartmann
@luraess Thanks for your informations. We would like to suggest **ranocha**, **simonbyrne**, and **ma-sadeghi** as reviewers. I just corrected the invalid DOIs. However, the two missing DOIs do not yet exist, as the papers are still in the review process. Is it necessary for all references to have a DOI?
