+++ test/discretization/test_multibody_setup.jl
@@ -7,33 +7,47 @@
     @test ms isa MultibodySetup{Peridynamics.BBMaterial{Peridynamics.NoCorrection},Peridynamics.BBPointParameters}
     @test Peridynamics.material_type(ms) == Peridynamics.BBMaterial{Peridynamics.NoCorrection}
     @test length(ms.bodies) == 2
-    @test isempty(ms.contacts)
+    @test isempty(ms.srf_contacts)
 
     ms = MultibodySetup(Dict(:body1 => b1, :body2 => b2))
     @test ms isa MultibodySetup{Peridynamics.BBMaterial{Peridynamics.NoCorrection},Peridynamics.BBPointParameters}
     @test Peridynamics.material_type(ms) == Peridynamics.BBMaterial{Peridynamics.NoCorrection}
     @test length(ms.bodies) == 2
-    @test isempty(ms.contacts)
+    @test isempty(ms.srf_contacts)
 
     @test_throws ArgumentError MultibodySetup(:body1 => b1, :body3 => b3)
 end
 
 @testitem "contact!" begin
+    using Peridynamics.PointNeighbors
+
     b1 = Body(BBMaterial(), rand(3,10), rand(10))
     b2 = Body(BBMaterial(), rand(3,10), rand(10))
     ms = MultibodySetup(:body1 => b1, :body2 => b2)
 
-    contact!(ms, :body1, :body2; radius=1)
-    @test ms.contacts[1] == Peridynamics.Contact(:body1, :body2, 1.0, 1e12)
-    empty!(ms.contacts)
+    nhs = GridNeighborhoodSearch{3}(1, 10)
 
-    contact!(ms, :body1, :body2; radius=2, sc=1.5e12)
-    @test ms.contacts[1] == Peridynamics.Contact(:body1, :body2, 2.0, 1.5e12)
-    empty!(ms.contacts)
+    contact!(ms, :body1, :body2; radius=1)
+    srfc = ms.srf_contacts[1] #Peridynamics.ShortRangeForceContact(:body1, :body2, 1.0, 1e12, nhs)
+    @test srfc.body_id_a === :body1
+    @test srfc.body_id_b === :body2
+    @test srfc.radius == 1.0
+    @test srfc.penalty_factor == 1e12
+    @test srfc.nhs isa GridNeighborhoodSearch{3}
+    empty!(ms.srf_contacts)
+
+    contact!(ms, :body1, :body2; radius=2, penalty_factor=1.5e12)
+    srfc = ms.srf_contacts[1] # Peridynamics.ShortRangeForceContact(:body1, :body2, 2.0, 1.5e12, nhs)
+    @test srfc.body_id_a === :body1
+    @test srfc.body_id_b === :body2
+    @test srfc.radius == 2.0
+    @test srfc.penalty_factor == 1.5e12
+    @test srfc.nhs isa GridNeighborhoodSearch{3}
+    empty!(ms.srf_contacts)
 
     @test_throws UndefKeywordError(:radius) contact!(ms, :body1, :body2)
 
     @test_throws ArgumentError contact!(ms, :body1, :body2; radius=0)
 
-    @test_throws ArgumentError contact!(ms, :body1, :body2; radius=1, sc=0)
+    @test_throws ArgumentError contact!(ms, :body1, :body2; radius=1, penalty_factor=0)
 end
