+++ src/Peridynamics.jl
@@ -77,7 +77,7 @@ include("discretization/point_generators.jl")
 include("discretization/predefined_cracks.jl")
 include("discretization/point_sets.jl")
 include("discretization/body.jl")
-include("discretization/contact.jl")
+include("discretization/short_range_force_contact.jl")
 include("discretization/multibody_setup.jl")
 include("discretization/decomposition.jl")
 include("discretization/chunk_handler.jl")
