### 2024-06-17 #150 @luraess
> @luraess Thanks for your informations. We would like to suggest **ranocha**, **simonbyrne**, and **ma-sadeghi** as reviewers. I just corrected the invalid DOIs. However, the two missing DOIs do not yet exist, as the papers are still in the review process. Is it necessary for all references to have a DOI?

@kaipartmann thanks for the update. For now let's keep these non resolving DOIs as such and we can fix things towards the end of the process, or see how to circumvent things if needed.

@ranocha @simonbyrne @ma-sadeghi - would you be willing to serve as reviewers for this 2 page contribution? Please 👍  or 👎 . Thanks.
