### 2024-06-18 1de4b6a
Add point generator for uniform spheres
