+++ src/Peridynamics.jl
@@ -77,7 +77,6 @@ include("discretization/point_generators.jl")
 include("discretization/predefined_cracks.jl")
 include("discretization/point_sets.jl")
 include("discretization/body.jl")
-include("discretization/short_range_force_contact.jl")
 include("discretization/multibody_setup.jl")
 include("discretization/decomposition.jl")
 include("discretization/chunk_handler.jl")
@@ -102,6 +101,7 @@ include("time_solvers/dynamic_relaxation.jl")
 include("physics/force_density.jl")
 include("physics/material_parameters.jl")
 include("physics/fracture.jl")
+include("physics/short_range_force_contact.jl")
 include("physics/bond_based.jl")
 include("physics/continuum_kinematics_inspired.jl")
 include("physics/ordinary_state_based.jl")
