### 2024-06-18 577bdfb
Add point set `:all_points` when creating a body
