### 2024-06-18 bfdad22
Move decomposition inside data handler construction
