+++ src/core/submit.jl
@@ -58,9 +58,8 @@ function submit_mpi(job::Job)
     @timeit_debug TO "initialization" begin
         init_logs(job.options)
         log_spatial_setup(job.options, job.spatial_setup)
-        point_decomp = PointDecomposition(job.spatial_setup, mpi_nranks())
         log_create_data_handler_start()
-        mdh = MPIDataHandler(job.spatial_setup, job.time_solver, point_decomp)
+        mdh = MPIDataHandler(job.spatial_setup, job.time_solver)
         init_time_solver!(job.time_solver, mdh)
         initialize!(mdh, job.time_solver)
         log_create_data_handler_end()
@@ -80,9 +79,8 @@ end
 function submit_threads(job::Job, n_chunks::Int)
     init_logs(job.options)
     log_spatial_setup(job.options, job.spatial_setup)
-    point_decomp = PointDecomposition(job.spatial_setup, n_chunks)
     log_create_data_handler_start()
-    tdh = ThreadsDataHandler(job.spatial_setup, job.time_solver, point_decomp)
+    tdh = ThreadsDataHandler(job.spatial_setup, job.time_solver, n_chunks)
     init_time_solver!(job.time_solver, tdh)
     initialize!(tdh, job.time_solver)
     log_create_data_handler_end()
