+++ src/core/submit.jl
@@ -59,7 +59,7 @@ function submit_mpi(job::Job)
         init_logs(job.options)
         log_spatial_setup(job.options, job.spatial_setup)
         log_create_data_handler_start()
-        mdh = MPIDataHandler(job.spatial_setup, job.time_solver)
+        mdh = MPIBodyDataHandler(job.spatial_setup, job.time_solver)
         init_time_solver!(job.time_solver, mdh)
         initialize!(mdh, job.time_solver)
         log_create_data_handler_end()
@@ -80,7 +80,7 @@ function submit_threads(job::Job, n_chunks::Int)
     init_logs(job.options)
     log_spatial_setup(job.options, job.spatial_setup)
     log_create_data_handler_start()
-    tdh = ThreadsDataHandler(job.spatial_setup, job.time_solver, n_chunks)
+    tdh = ThreadsBodyDataHandler(job.spatial_setup, job.time_solver, n_chunks)
     init_time_solver!(job.time_solver, tdh)
     initialize!(tdh, job.time_solver)
     log_create_data_handler_end()
