+++ src/Peridynamics.jl
@@ -93,8 +93,8 @@ include("core/materials.jl")
 include("core/storages.jl")
 include("core/time_solvers.jl")
 include("core/halo_exchange.jl")
-include("core/threads_data_handler.jl")
-include("core/mpi_data_handler.jl")
+include("core/threads_body_data_handler.jl")
+include("core/mpi_body_data_handler.jl")
 
 include("time_solvers/velocity_verlet.jl")
 include("time_solvers/dynamic_relaxation.jl")
