### 2024-06-18 7ecf27e
Created special constructor functions
