+++ src/core/data_handler.jl
@@ -1,0 +1,9 @@
+function initialize!(dh, solver)
+    return nothing
+end
+
+function log_data_handler end
+
+function threads_data_handler end
+
+function mpi_data_handler end
