+++ test/core/test_halo_exchange.jl
@@ -115,9 +115,9 @@ end
     forcedensity_bc!(t -> t, body, :a, :x)
     precrack!(body, :a, :b)
     ts = VelocityVerlet(steps=10)
-    tdh = Peridynamics.ThreadsBodyDataHandler(body, ts, 2)
+    dh = Peridynamics.threads_data_handler(body, ts, 2)
 
-    b1 = tdh.chunks[1]
+    b1 = dh.chunks[1]
     @test b1 isa Peridynamics.BodyChunk
     @test b1.storage.position == position
     @test b1.storage.displacement == zeros(3, 2)
@@ -129,7 +129,7 @@ end
     @test b1.storage.damage ≈ [2/3, 2/3]
     @test b1.storage.bond_active == [1, 0, 0, 1, 0, 0]
     @test b1.storage.n_active_bonds == [1, 1]
-    b2 = tdh.chunks[2]
+    b2 = dh.chunks[2]
     @test b2 isa Peridynamics.BodyChunk
     @test b2.storage.position == position[:, [3, 4, 1, 2]]
     @test b2.storage.displacement == zeros(3, 2)
@@ -143,8 +143,8 @@ end
     @test b2.storage.n_active_bonds == [1, 1]
 
     randpos = rand(3, 4)
-    tdh.chunks[2].storage.position .= randpos
-    Peridynamics.exchange_loc_to_halo!(tdh, 1)
+    dh.chunks[2].storage.position .= randpos
+    Peridynamics.exchange_loc_to_halo!(dh, 1)
 
     @test b1 isa Peridynamics.BodyChunk
     @test b1.storage.position[:,1:2] ≈ [0.0 1.0; 0.0 0.0; 0.0 0.0]
