+++ src/core/data_handler.jl
@@ -2,8 +2,14 @@ function initialize!(dh, solver)
     return nothing
 end
 
-function log_data_handler end
+function log_data_handler(options, dh)
+    return nothing
+end
 
-function threads_data_handler end
+function threads_data_handler(ms, solver, n_chunks)
+    return nothing
+end
 
-function mpi_data_handler end
+function mpi_data_handler(body, solver)
+    return nothing
+end
