+++ src/Peridynamics.jl
@@ -99,7 +99,9 @@ include("core/time_solvers.jl")
 include("core/halo_exchange.jl")
 include("core/data_handler.jl")
 include("core/threads_body_data_handler.jl")
+include("core/threads_multibody_data_handler.jl")
 include("core/mpi_body_data_handler.jl")
+include("core/mpi_multibody_data_handler.jl")
 
 include("time_solvers/velocity_verlet.jl")
 include("time_solvers/dynamic_relaxation.jl")
