+++ src/auxiliary/io.jl
@@ -28,9 +28,9 @@ function check_export_fields(::Type{S}, fields::Vector{Symbol}) where {S}
     return nothing
 end
 
-function _export_results(b::AbstractBodyChunk, chunk_id::Int, n_chunks::Int,
-                         options::AbstractOptions, n::Int, t::Float64)
-    filename = joinpath(options.vtk, @sprintf("timestep_%06d", n))
+function _export_results(options::AbstractOptions, b::AbstractBodyChunk, chunk_id::Int,
+                         n_chunks::Int, prefix::AbstractString, n::Int, t::Float64)
+    filename = joinpath(options.vtk, @sprintf("%s_timestep_%06d", prefix, n))
     position = get_loc_position(b)
     pvtk_grid(filename, position, b.cells; part=chunk_id, nparts=n_chunks) do vtk
         for field in options.fields
