### 2024-06-19 54dc4bd
Refactoring `contact!`
