+++ src/discretization/multibody_setup.jl
@@ -90,47 +90,6 @@ end
 @inline each_body_idx(ms::AbstractMultibodySetup) = eachindex(ms.bodies)
 @inline each_body_name(ms::AbstractMultibodySetup) = ms.body_names
 
-"""
-    contact!(ms, body_a, body_b; kwargs...)
-
-Defines contact between multiple bodies
-
-# Arguments
-
-- `ms::AbstractMultibodySetup`: The multibody setup defined to simulate the contact
-- `body_a::Symbol`: First body in contact
-- `body_b::Symbol`: Second body in contact
-
-# Keywords
-
-- `radius::Float64`:
-- `sc::Float64`:
-
-# Throws
-
-- Error if a called body is not defined in the multibody setup
-- Error if keyword is not allowed
-
-TODO kwargs
-"""
-function contact!(ms::AbstractMultibodySetup, name_body_a::Symbol, name_body_b::Symbol;
-                  kwargs...)
-    check_if_bodyname_is_defined(ms, name_body_a)
-    check_if_bodyname_is_defined(ms, name_body_b)
-
-    p = Dict{Symbol,Any}(kwargs)
-    check_kwargs(p, CONTACT_KWARGS)
-    radius, penalty_factor = get_contact_params(p)
-
-    body_b = get_body(ms, name_body_b)
-    nhs = GridNeighborhoodSearch{3}(radius, body_b.n_points)
-
-    srfc = ShortRangeForceContact(name_body_a, name_body_b, radius, penalty_factor, nhs)
-    push!(ms.srf_contacts, srfc)
-
-    return nothing
-end
-
 function pre_submission_check(ms::AbstractMultibodySetup)
     #TODO: check if everything is defined for job submission!
     return nothing
