+++ src/Peridynamics.jl
@@ -77,11 +77,15 @@ include("conditions/boundary_conditions.jl")
 include("conditions/initial_conditions.jl")
 include("conditions/condition_checks.jl")
 
+include("physics/force_density.jl")
+include("physics/material_parameters.jl")
+include("physics/fracture.jl")
+include("physics/short_range_force_contact.jl")
+
 include("discretization/point_generators.jl")
 include("discretization/predefined_cracks.jl")
 include("discretization/point_sets.jl")
 include("discretization/body.jl")
-include("discretization/short_range_force_contact.jl")
 include("discretization/multibody_setup.jl")
 include("discretization/decomposition.jl")
 include("discretization/chunk_handler.jl")
@@ -106,9 +110,6 @@ include("core/mpi_multibody_data_handler.jl")
 include("time_solvers/velocity_verlet.jl")
 include("time_solvers/dynamic_relaxation.jl")
 
-include("physics/force_density.jl")
-include("physics/material_parameters.jl")
-include("physics/fracture.jl")
 include("physics/bond_based.jl")
 include("physics/continuum_kinematics_inspired.jl")
 include("physics/ordinary_state_based.jl")
