### 2024-06-20 ec3d54b
Add functions for initializing the NHS
