### 2024-06-20 9a2984e
MultibodySetup can now hold bodies with arbitrary materials
