### 2024-06-20 09af85b
Clean error for using MPI on multibody simulations
