+++ src/core/mpi_multibody_data_handler.jl
@@ -1,3 +1,3 @@
 function mpi_data_handler(::AbstractMultibodySetup, ::AbstractTimeSolver)
-    return error("Multibody simulations with MPI not yet implemented!\n")
+    return nothing
 end
