### 2024-06-20 2c60301
Simulations work if `nthreads` > `n_points`
