+++ src/core/mpi_multibody_data_handler.jl
@@ -1,3 +1,4 @@
-function mpi_data_handler(::AbstractMultibodySetup, ::AbstractTimeSolver)
-    return nothing
-end
+# TODO
+# function mpi_data_handler(::AbstractMultibodySetup, ::AbstractTimeSolver)
+#     return nothing
+# end
