+++ src/core/threads_multibody_data_handler.jl
@@ -63,8 +63,8 @@ function calc_contact_force_densities!(dh)
     return nothing
 end
 
-function export_reference_results(dh::ThreadsMultibodyDataHandler, options::AbstractOptions)
-    options.exportflag || return nothing
+function export_reference_results(dh::ThreadsMultibodyDataHandler, options::AbstractJobOptions)
+    options.export_allowed || return nothing
     for body_idx in each_body_idx(dh)
         body_dh = get_body_dh(dh, body_idx)
         name = get_body_name(dh, body_idx)

[…]
