### 2024-06-21 f9f3a85
Made `JobOptions` typestable
