### 2024-06-21 ee4be9e
Small fixes and improved tests for io
