+++ test/auxiliary/test_io.jl
@@ -1,54 +1,225 @@
-@testitem "JobOptions BBVerletStorage" begin
-    o = Dict{Symbol,Any}(:path => "rootpath", :freq => 10)
-    eo = Peridynamics.get_job_options(Peridynamics.BBVerletStorage, o)
-    @test eo.export_allowed == true
-    @test eo.root == "rootpath"
-    @test eo.vtk == joinpath("rootpath", "vtk")
-    @test eo.logfile == joinpath("rootpath", "logfile.log")
-    @test eo.freq == 10
-    @test eo.fields == [field for field in Peridynamics.DEFAULT_EXPORT_FIELDS]
+@testitem "JobOptions" begin
+    function tests_body_specific(body::Body, solver)
+        default_fields_body = [field for field in Peridynamics.default_export_fields()]
 
-    o = Dict{Symbol,Any}(:path => "rootpath")
-    eo = Peridynamics.get_job_options(Peridynamics.BBVerletStorage, o)
-    @test eo.export_allowed == true
-    @test eo.root == "rootpath"
-    @test eo.vtk == joinpath("rootpath", "vtk")
-    @test eo.logfile == joinpath("rootpath", "logfile.log")
-    @test eo.freq == 10
-    @test eo.fields == [field for field in Peridynamics.DEFAULT_EXPORT_FIELDS]
+        o = Dict{Symbol,Any}(:path => "rootpath", :freq => 10)
+        jo = Peridynamics.get_job_options(body, solver, o)
+        @test jo.export_allowed == true
+        @test jo.root == "rootpath"
+        @test jo.vtk == joinpath("rootpath", "vtk")
+        @test jo.logfile == joinpath("rootpath", "logfile.log")
+        @test jo.freq == 10
+        @test jo.fields == default_fields_body
 
-    o = Dict{Symbol,Any}(:freq => 10)
-    msg = "if `freq` is specified, the keyword `path` is also needed!\n"
-    @test_throws ArgumentError(msg) begin
-        Peridynamics.get_job_options(Peridynamics.BBVerletStorage, o)
+        o = Dict{Symbol,Any}(:path => "rootpath")
+        jo = Peridynamics.get_job_options(body, solver, o)
+        @test jo.export_allowed == true
+        @test jo.root == "rootpath"
+        @test jo.vtk == joinpath("rootpath", "vtk")
+        @test jo.logfile == joinpath("rootpath", "logfile.log")
+        @test jo.freq == 10
+        @test jo.fields == default_fields_body
+
+        o = Dict{Symbol,Any}(:freq => 10)
+        msg = "if `freq` is specified, the keyword `path` is also needed!\n"
+        @test_throws ArgumentError(msg) begin
+            Peridynamics.get_job_options(body, solver, o)
+        end
+
+        o = Dict{Symbol,Any}()
+        jo = Peridynamics.get_job_options(body, solver, o)
+        @test jo.export_allowed == false
+        @test jo.root == ""
+        @test jo.vtk == ""
+        @test jo.logfile == ""
+        @test jo.freq == 0
+        @test jo.fields == Vector{Symbol}()
+
+        o = Dict{Symbol,Any}(:path => "rootpath", :freq => -10)
+        msg = "`freq` should be larger than zero!\n"
+        @test_throws ArgumentError(msg) begin
+            Peridynamics.get_job_options(body, solver, o)
+        end
+
+        o = Dict{Symbol,Any}(:path => "rootpath", :fields => (:displacement, :b_ext))
+        jo = Peridynamics.get_job_options(body, solver, o)
+        @test jo.export_allowed == true
+        @test jo.root == "rootpath"
+        @test jo.vtk == joinpath("rootpath", "vtk")
[…]
