### 2024-06-21 ab0a6ab
Failed attempt to make `MultibodySetup` typestable
