### 2024-06-21 7ee026a
Use serial `update_grid!` instead of `initialize_grid!`
