+++ src/discretization/bond_system_corrections.jl
@@ -49,10 +49,10 @@ end
 
 function initialize!(dh::AbstractThreadsBodyDataHandler{BondSystem{EnergySurfaceCorrection}},
                      ::AbstractTimeSolver)
-    @threads :static for chunk in dh.chunks
+    @batch for chunk in dh.chunks
         calc_mfactor!(chunk)
     end
-    @threads :static for chunk_id in eachindex(dh.chunks)
+    @batch for chunk_id in eachindex(dh.chunks)
         exchange_loc_to_halo!(get_mfactor, dh, chunk_id)
         calc_scfactor!(dh.chunks[chunk_id])
     end
