### 2024-06-21 77268ce
Extend show method for point parameters
