### 2024-06-24 9a7ec2f
Renaming `log_qty` -> `msg_qty` and added separator
