+++ src/physics/material_parameters.jl
@@ -128,12 +128,12 @@ required_point_parameters() = (:δ, :rho, :E, :nu, :G, :K, :λ, :μ, :Gc, :εc)
 # req_param_material(::AbstractMaterial) = (:δ, :rho, :E, :nu, :G, :K, :λ, :μ)
 
 function log_material_parameters(param::AbstractPointParameters; indentation::Int=2)
-    msg = log_qty("horizon", param.δ; indentation=indentation)
-    msg *= log_qty("density", param.rho; indentation=indentation)
-    msg *= log_qty("Young's modulus", param.E; indentation=indentation)
-    msg *= log_qty("Poisson's ratio", param.nu; indentation=indentation)
-    msg *= log_qty("shear modulus", param.G; indentation=indentation)
-    msg *= log_qty("bulk modulus", param.K; indentation=indentation)
+    msg = msg_qty("horizon", param.δ; indentation=indentation)
+    msg *= msg_qty("density", param.rho; indentation=indentation)
+    msg *= msg_qty("Young's modulus", param.E; indentation=indentation)
+    msg *= msg_qty("Poisson's ratio", param.nu; indentation=indentation)
+    msg *= msg_qty("shear modulus", param.G; indentation=indentation)
+    msg *= msg_qty("bulk modulus", param.K; indentation=indentation)
     return msg
 end
 
