### 2024-06-24 48dd137
Added show methods for all important public types
