+++ src/time_solvers/velocity_verlet.jl
@@ -69,6 +69,36 @@ mutable struct VelocityVerlet <: AbstractTimeSolver
     end
 end
 
+function Base.show(io::IO, vv::VelocityVerlet)
+    print(io, typeof(vv))
+    fields = Vector{Symbol}()
+    for field in fieldnames(typeof(vv))
+        value = getfield(vv, field)
+        if value > 0
+            push!(fields, field)
+        end
+    end
+    print(io, msg_fields_in_brackets(vv, Tuple(fields)))
+    return nothing
+end
+
+function Base.show(io::IO, ::MIME"text/plain", vv::VelocityVerlet)
+    if get(io, :compact, false)
+        Base.show_default(io, vv)
+    else
+        println(io, typeof(vv), ":")
+        fields = Vector{Symbol}()
+        for field in fieldnames(typeof(vv))
+            value = getfield(vv, field)
+            if value > 0
+                push!(fields, field)
+            end
+        end
+        print(io, msg_fields(vv, Tuple(fields)))
+    end
+    return nothing
+end
+
 function init_time_solver!(vv::VelocityVerlet, dh::AbstractDataHandler)
     if vv.Δt < 0
         vv.Δt = calc_stable_timestep(dh, vv.safety_factor)
