### 2024-06-24 f7ad996
Changed show method for compact io context
