### 2024-06-24 a386c83
Added tests for show methods
