+++ test/discretization/test_body.jl
@@ -412,6 +412,30 @@ end
     @test contains(msg, "2-point set `a`")
     @test contains(msg, "10-point set `all_points`")
 
+    material!(body, horizon=1, rho=1, E=1, Gc=1)
+
+    show(IOContext(io, :compact=>true), MIME("text/plain"), body)
+    msg = String(take!(io))
+    @test msg == "10-point Body{BBMaterial{NoCorrection}}"
+
+    show(IOContext(io, :compact=>false), MIME("text/plain"), body)
+    msg = String(take!(io))
+    @test contains(msg, "1 point parameter(s):")
+    @test contains(msg, "BBPointParameters(δ=1.0, E=1.0, nu=0.25, rho=1.0, Gc=1.0)")
+
+    material!(body, :a, horizon=2, rho=2, E=2, Gc=2)
+
+    show(IOContext(io, :compact=>true), MIME("text/plain"), body)
+    msg = String(take!(io))
+    @test msg == "10-point Body{BBMaterial{NoCorrection}}"
+
+    show(IOContext(io, :compact=>false), MIME("text/plain"), body)
+    msg = String(take!(io))
+    @test !contains(msg, "1 point parameter(s):")
+    @test contains(msg, "2 point parameter(s):")
+    @test contains(msg, "BBPointParameters(δ=1.0, E=1.0, nu=0.25, rho=1.0, Gc=1.0)")
+    @test contains(msg, "BBPointParameters(δ=2.0, E=2.0, nu=0.25, rho=2.0, Gc=2.0)")
+
     velocity_ic!(body, :a, :z, 1.0)
     velocity_bc!(t -> t, body, :a, 1)
     forcedensity_bc!((p, t) -> p[1] + p[2] + p[3] + t, body, :a, 2)
@@ -434,6 +458,18 @@ end
     @test contains(msg, "SingleDimBC(field=velocity_half, point_set=a, dim=1)")
     @test contains(msg, "1 initial condition(s):")
     @test contains(msg, "SingleDimIC(value=1.0, field=velocity, point_set=a, dim=3)")
+
+    point_set!(body, :b, 3:4)
+
+    precrack!(body, :a, :b)
+
+    show(IOContext(io, :compact=>true), MIME("text/plain"), body)
+    msg = String(take!(io))
+    @test msg == "10-point Body{BBMaterial{NoCorrection}}"
+
+    show(IOContext(io, :compact=>false), MIME("text/plain"), body)
+    msg = String(take!(io))
+    @test contains(msg, "1 predefined crack(s)")
 
     Peridynamics.change_name!(body, :testbody)
 
