+++ src/Peridynamics.jl
@@ -69,6 +69,7 @@ abstract type AbstractCorrection end
 abstract type AbstractStorage end
 abstract type AbstractCondition end
 abstract type AbstractBondSystemMaterial{Correction} <: AbstractMaterial end
+abstract type AbstractInteractionSystemMaterial <: AbstractMaterial end
 
 include("auxiliary/function_arguments.jl")
 include("auxiliary/io.jl")
@@ -93,6 +94,7 @@ include("discretization/decomposition.jl")
 include("discretization/chunk_handler.jl")
 include("discretization/bond_system.jl")
 include("discretization/bond_system_corrections.jl")
+include("discretization/interaction_system.jl")
 include("discretization/body_chunk.jl")
 
 include("core/job.jl")
