### 2024-06-26 58636ad
Add search algorithm for two- and three-NI
