+++ src/discretization/interaction_system.jl
@@ -28,19 +28,25 @@ end
 
 function InteractionSystem(body::AbstractBody, pd::PointDecomposition, chunk_id::Int)
     check_interaction_system_compat(body.mat)
-    bonds, n_one_nis = find_bonds(body, pd.decomp[chunk_id])
+    loc_points = pd.decomp[chunk_id]
+    bonds, n_one_nis = find_bonds(body, loc_points)
     bond_ids = find_bond_ids(n_one_nis)
+    volume_one_nis = zeros(length(n_one_nis))
     if has_two_nis(body)
-        two_nis, n_two_nis = find_two_nis(body.position, bonds, bond_ids)
+        two_nis, n_two_nis = find_two_nis(body, loc_points, bonds, bond_ids)
+        volume_two_nis = zeros(length(n_two_nis))
     else
         two_nis = Vector{TwoNeighborInteraction}()
         n_two_nis = Vector{Float64}()
+        volume_two_nis = Vector{Float64}()
     end
     if has_two_nis(body)
-        three_nis, n_three_nis = find_three_nis(body.position, bonds, bond_ids)
+        three_nis, n_three_nis = find_three_nis(body, loc_points, bonds, bond_ids)
+        volume_three_nis = zeros(length(n_three_nis))
     else
         three_nis = Vector{ThreeNeighborInteraction}()
         n_three_nis = Vector{Float64}()
+        volume_three_nis = Vector{Float64}()
     end
     ch = get_chunk_handler(bonds, pd, chunk_id)
     localize!(bonds, ch.localizer)
@@ -77,9 +83,130 @@ function has_two_nis(body::AbstractBody)
     return false
 end
 
+@inline function has_two_nis(body::AbstractBody, point_id::Int)
+    get_c2(body.point_params[body.params_map[point_id]]) ≈ 0 || return true
+    return false
+end
+
 function has_three_nis(body::AbstractBody)
     for params in body.point_params
         get_c3(params) ≈ 0 || return true
     end
     return false
 end
+
+@inline function has_three_nis(body::AbstractBody, point_id::Int)
+    get_c3(body.point_params[body.params_map[point_id]]) ≈ 0 || return true
+    return false
+end
+
+function find_two_nis(body, loc_points, bonds, bond_ids)
+    two_nis = Vector{TwoNeighborInteraction}()
+    sizehint!(two_nis, n_points(body) * 1000)
+    n_two_nis = zeros(length(loc_points))
+    position = body.position
+    points_with_two_nis = filter(x -> has_two_nis(body, x), loc_points)
+    for i in points_with_two_nis
+        num = 0
+        δ = get_point_param(body, :δ, i)
+        jk_seen = Set{Tuple{Int,Int}}()
+        for oni_j in bond_ids[i], oni_k in bond_ids[i]
+            j, k = bonds[oni_j].neighbor, bonds[oni_k].neighbor
+            if k !== j && !in((j, k), ijseen)
+                Ξijx = position[1, j] - position[1, i]
+                Ξijy = position[2, j] - position[2, i]
+                Ξijz = position[3, j] - position[3, i]
+                Ξikx = position[1, k] - position[1, i]
+                Ξiky = position[2, k] - position[2, i]
+                Ξikz = position[3, k] - position[3, i]
+                Ξjkx = position[1, k] - position[1, j]
+                Ξjky = position[2, k] - position[2, j]
+                Ξjkz = position[3, k] - position[3, j]
+                Ξjk = sqrt(Ξjkx * Ξjkx + Ξjky * Ξjky + Ξjkz * Ξjkz)
[…]
