+++ src/discretization/body_chunk.jl
@@ -45,6 +45,8 @@ end
 @inline each_point_idx(b::AbstractBodyChunk) = each_point_idx(b.ch)
 @inline each_point_idx_pair(b::AbstractBodyChunk) = each_point_idx_pair(b.ch)
 
+@inline n_loc_points(b::AbstractBodyChunk) = length(b.ch.loc_points)
+
 function initialize!(::BodyChunk)
     return nothing
 end
