### 2024-06-27 69c1d41
Add tests for two- and three-NIs
