+++ src/physics/continuum_kinematics_inspired.jl
@@ -39,10 +39,44 @@ function CKIPointParameters(::CKIMaterial, p::Dict{Symbol,Any})
     rho = get_density(p)
     E, nu, G, K, λ, μ = get_elastic_params(p)
     Gc, εc = get_frac_params(p, δ, K)
-    C1 = 30 / π * μ / δ^4
-    C2 = 0.0
-    C3 = 32 / π^4 * (λ - μ) / δ^12
+    C1, C2, C3 = cki_parameters(p, δ, λ, μ)
     return CKIPointParameters(δ, rho, E, nu, G, K, λ, μ, Gc, εc, C1, C2, C3)
+end
+
+function cki_parameters(p::Dict{Symbol,Any}, δ, λ, μ)
+    if haskey(p, :C1)
+        C1::Float64 = float(p[:C1])
+    else
+        C1 = 0.0
+    end
+
+    if haskey(p, :C2)
+        C2::Float64 = float(p[:C2])
+    else
+        C2 = 0.0
+    end
+
+    if haskey(p, :C3)
+        C3::Float64 = float(p[:C3])
+    else
+        C3 = 0.0
+    end
+
+    if C1 ≈ 0 && C2 ≈ 0 && C3 ≈ 0
+        C1 = 30 / π * μ / δ^4
+        C2 = 0.0
+        C3 = 32 / π^4 * (λ - μ) / δ^12
+    else
+        msg = "parameters for CKIMaterial specified manually!\n"
+        msg *= "Be careful when adjusting these parameters to avoid unexpected outcomes!"
+        @warn msg
+    end
+
+    return C1, C2, C3
+end
+
+function allowed_material_kwargs(::CKIMaterial)
+    return (DEFAULT_POINT_KWARGS..., :C1, :C2, :C3)
 end
 
 @params CKIMaterial CKIPointParameters
