+++ test/discretization/test_interaction_system.jl
@@ -1,0 +1,76 @@
+@testitem "one- and two-neighbor interactions" begin
+    position = [0.0 1.0 0.0
+                0.0 0.0 0.0
+                0.0 0.0 1.0]
+    volume = fill(1.0, 3)
+    body = Body(CKIMaterial(), position, volume)
+    material!(body; horizon=1.5, rho=8e-6, E=210e3, nu=0.3, Gc=1.0, C1=1, C2=1)
+    pd = Peridynamics.PointDecomposition(body, 1)
+
+    is, ch = Peridynamics.InteractionSystem(body, pd, 1)
+
+    @test is.one_nis == [
+        Peridynamics.Bond(2, 1.0, true),
+        Peridynamics.Bond(3, 1.0, true),
+        Peridynamics.Bond(1, 1.0, true),
+        Peridynamics.Bond(3, √2, true),
+        Peridynamics.Bond(1, 1.0, true),
+        Peridynamics.Bond(2, √2, true)]
+    @test is.n_one_nis == [2, 2, 2]
+    @test is.bond_ids == [1:2, 3:4, 5:6]
+    @test is.two_nis == [
+        Peridynamics.TwoNeighborInteraction(1, 2, 1.0),
+        Peridynamics.TwoNeighborInteraction(3, 4, 1.0),
+        Peridynamics.TwoNeighborInteraction(5, 6, 1.0)]
+    @test is.n_two_nis == [1, 1, 1]
+    @test is.three_nis == Vector{Peridynamics.ThreeNeighborInteraction}()
+    @test is.n_three_nis == Vector{Int}()
+end
+
+@testitem "one-, two-, and three-neighbor interactions" begin
+    position = [0.0 1.0 0.0 0.0 2.0
+                0.0 0.0 1.0 0.0 2.0
+                0.0 0.0 0.0 1.0 2.0]
+    volume = fill(1.0, 5)
+    body = Body(CKIMaterial(), position, volume)
+    material!(body; horizon=1.5, rho=8e-6, E=210e3, nu=0.3, Gc=1.0, C1=1, C2=1, C3=1)
+    pd = Peridynamics.PointDecomposition(body, 1)
+
+    is, ch = Peridynamics.InteractionSystem(body, pd, 1)
+
+    @test is.one_nis == [
+        Peridynamics.Bond(2, 1.0, true),
+        Peridynamics.Bond(3, 1.0, true),
+        Peridynamics.Bond(4, 1.0, true),
+        Peridynamics.Bond(1, 1.0, true),
+        Peridynamics.Bond(3, √2, true),
+        Peridynamics.Bond(4, √2, true),
+        Peridynamics.Bond(1, 1.0, true),
+        Peridynamics.Bond(2, √2, true),
+        Peridynamics.Bond(4, √2, true),
+        Peridynamics.Bond(1, 1.0, true),
+        Peridynamics.Bond(2, √2, true),
+        Peridynamics.Bond(3, √2, true)]
+    @test is.n_one_nis == [3, 3, 3, 3, 0]
+    @test is.bond_ids == [1:3, 4:6, 7:9, 10:12, 13:12]
+    @test is.two_nis == [
+        Peridynamics.TwoNeighborInteraction(1, 2, 1.0),
+        Peridynamics.TwoNeighborInteraction(1, 3, 1.0),
+        Peridynamics.TwoNeighborInteraction(2, 3, 1.0),
+        Peridynamics.TwoNeighborInteraction(4, 5, 1.0),
+        Peridynamics.TwoNeighborInteraction(4, 6, 1.0),
+        Peridynamics.TwoNeighborInteraction(5, 6, √3),
[…]
