### 2024-06-27 de29a84
added 'TODO' docstrings to all exported functions for #49
