+++ src/physics/ordinary_state_based.jl
@@ -46,6 +46,9 @@ struct OSBMaterial{Correction} <: AbstractBondSystemMaterial{Correction} end
 
 OSBMaterial() = OSBMaterial{NoCorrection}()
 
+"""
+TODO
+"""
 struct OSBPointParameters <: AbstractPointParameters
     δ::Float64
     rho::Float64
