### 2024-06-28 951dbbb
CKIMaterial working as in `v0.2`
