+++ src/Peridynamics.jl
@@ -80,7 +80,6 @@ include("conditions/boundary_conditions.jl")
 include("conditions/initial_conditions.jl")
 include("conditions/condition_checks.jl")
 
-include("physics/force_density.jl")
 include("physics/material_parameters.jl")
 include("physics/fracture.jl")
 include("physics/short_range_force_contact.jl")
