+++ src/time_solvers/dynamic_relaxation.jl
@@ -117,7 +117,21 @@ end
     return nothing
 end
 
-function relaxation_timestep!(dh::AbstractThreadsBodyDataHandler, options::AbstractJobOptions,
+@inline function calc_density_matrix!(density_matrix::Matrix{Float64},
+                                      system::InteractionSystem,
+                                      params::AbstractPointParameters,
+                                      dr::DynamicRelaxation, i::Int)
+    n_one_nis = system.n_one_nis[i]
+    k = 5π * params.δ^2 * params.C1
+    Λ = dr.Λ * 1 / 4 * dr.Δt^2 * n_one_nis * k
+    density_matrix[1, i] = Λ
+    density_matrix[2, i] = Λ
+    density_matrix[3, i] = Λ
+    return nothing
+end
+
+function relaxation_timestep!(dh::AbstractThreadsBodyDataHandler,
+                              options::AbstractJobOptions,
                               Δt::Float64, n::Int)
     t = n * Δt
     @batch for chunk_id in eachindex(dh.chunks)

[…]
