+++ src/discretization/interaction_system.jl
@@ -347,8 +347,7 @@ function calc_force_density!(chunk::AbstractBodyChunk{S,M}) where {S<:Interactio
     storage.b_int .= 0
     storage.n_active_one_nis .= 0
     for point_id in each_point_idx(chunk)
-        params = get_params(paramsetup, point_id)
-        force_density_point!(storage, system, mat, params, point_id)
+        force_density_point!(storage, system, mat, paramsetup, point_id)
     end
     return nothing
 end
