+++ src/physics/continuum_kinematics_inspired.jl
@@ -201,6 +201,40 @@ function force_density_point_one_ni!(storage::CKIStorage, system::InteractionSys
     return nothing
 end
 
+function force_density_point_one_ni!(storage::CKIStorage, system::InteractionSystem,
+                                     ::CKIMaterial, paramhandler::ParameterHandler, i::Int)
+    params_i = get_params(paramhandler, i)
+    for one_ni_id in each_one_ni_idx(system, i)
+        bond = system.one_nis[one_ni_id]
+        j, L = bond.neighbor, bond.length
+
+        # current bond length
+        Δxijx = storage.position[1, j] - storage.position[1, i]
+        Δxijy = storage.position[2, j] - storage.position[2, i]
+        Δxijz = storage.position[3, j] - storage.position[3, i]
+        l = sqrt(Δxijx * Δxijx + Δxijy * Δxijy + Δxijz * Δxijz)
+
+        # bond strain
+        ε = (l - L) / L
+
+        # failure mechanism
+        if ε > params.εc && bond.fail_permit
+            storage.one_ni_active[one_ni_id] = false
+        end
+        storage.n_active_one_nis[i] += storage.one_ni_active[one_ni_id]
+
+        # update of force density
+        one_ni_fail = storage.one_ni_active[one_ni_id]
+        params_j = get_params(paramhandler, j)
+        C1_effective = (params_i.C1 + params_j.C1) / 2
+        temp = one_ni_fail * C1_effective * (1 / L - 1 / l) * system.volume_one_nis[i]
+        storage.b_int[1, i] += temp * Δxijx
+        storage.b_int[2, i] += temp * Δxijy
+        storage.b_int[3, i] += temp * Δxijz
+    end
+    return nothing
+end
+
 function force_density_point_two_ni!(storage::CKIStorage, system::InteractionSystem,
                                      ::CKIMaterial, params::CKIPointParameters, i::Int)
     for two_ni_id in each_two_ni_idx(system, i)
@@ -229,11 +263,52 @@ function force_density_point_two_ni!(storage::CKIStorage, system::InteractionSys
         storage.b_int[3, i] += temp * (Δxikx * aijky - Δxiky * aijkx)
         # variable switch: i,j = j,i
         storage.b_int[1, i] += temp * (Δxijy * (Δxikx * Δxijy - Δxiky * Δxijx) -
-                                       Δxijz * (Δxikz * Δxijx - Δxikx * Δxijz))
+                                Δxijz * (Δxikz * Δxijx - Δxikx * Δxijz))
         storage.b_int[2, i] += temp * (Δxijz * (Δxiky * Δxijz - Δxikz * Δxijy) -
-                                       Δxijx * (Δxikx * Δxijy - Δxiky * Δxijx))
+                                Δxijx * (Δxikx * Δxijy - Δxiky * Δxijx))
         storage.b_int[3, i] += temp * (Δxijx * (Δxikz * Δxijx - Δxikx * Δxijz) -
-                                       Δxijy * (Δxiky * Δxijz - Δxikz * Δxijy))
+                                Δxijy * (Δxiky * Δxijz - Δxikz * Δxijy))
+    end
+    return nothing
+end
+
+function force_density_point_two_ni!(storage::CKIStorage, system::InteractionSystem,
+                                     ::CKIMaterial, paramhandler::ParameterHandler, i::Int)
+    params_i = get_params(paramhandler, i)
+    for two_ni_id in each_two_ni_idx(system, i)
[…]
