### 2024-06-28 864379e
Code enhancements in force density calculation
