+++ src/discretization/bond_system.jl
@@ -204,6 +204,16 @@ end
     return nothing
 end
 
+@inline function stretch_based_failure!(storage::AbstractStorage, ::BondSystem,
+                                        bond::Bond, params::AbstractPointParameters,
+                                        ε::Float64, i::Int, bond_id::Int)
+    if ε > params.εc && bond.fail_permit
+        storage.bond_active[bond_id] = false
+    end
+    storage.n_active_bonds[i] += storage.bond_active[bond_id]
+    return nothing
+end
+
 function log_system(::Type{B}, options::AbstractJobOptions,
                     dh::AbstractDataHandler) where {B<:BondSystem}
     n_bonds = calc_n_bonds(dh)
