### 2024-06-29 7964870
CompatHelper: bump compat for PointNeighbors to 0.4, (keep existing compat)
