+++ Project.toml
@@ -31,7 +31,7 @@ LibGit2 = "^1.8"
 LightXML = "^0.9.0"
 LinearAlgebra = "^1.8"
 MPI = "~0.20.0"
-PointNeighbors = "^0.3.0"
+PointNeighbors = "^0.3.0, 0.4"
 Polyester = "^0.7.5"
 PrecompileTools = "^1.0.0"
 Printf = "^1.8"
