+++ docs/make.jl
@@ -32,6 +32,7 @@ LIT_TUTORIALS_IN = [
     "tutorial_tension_dynfrac.jl",
     "tutorial_tension_precrack.jl",
     "tutorial_wave_in_bar.jl",
+    "tutorial_kalthoff-winkler_dynfrac.jl",
 ]
 LIT_TUTORIALS_IN .= joinpath.(@__DIR__, "src", "literate", LIT_TUTORIALS_IN)
 Literate.markdown.(LIT_TUTORIALS_IN, LIT_MD_OUT; credit=false)
@@ -76,6 +77,7 @@ makedocs(;
             joinpath("generated", "tutorial_tension_dynfrac.md"),
             joinpath("generated", "tutorial_tension_precrack.md"),
             joinpath("generated", "tutorial_wave_in_bar.md"),
+            joinpath("generated", "tutorial_kalthoff-winkler_dynfrac.md"),
             # "tutorial_logo.md",
         ],
         "API" => [
