+++ docs/src/literate/tutorial_kalthoff-winkler_dynfrac.jl
@@ -1,23 +1,43 @@
 # # [Kalthoff-Winkler experiment](@id tutorial_kalthoff-winkler)
-
-# Import the package:
-using Peridynamics
-using GLMakie #src
-using LaTeXStrings #src
-
+#
 # This tutorial demonstrates how to set up and run the Kalthoff-Winkler experiment using the Peridynamics.jl package.
-# The Kalthoff-Winkler experiment is a classic dynamic fracture experiment involving a pre-notched sample subjected to an impact loading.
+# The Kalthoff-Winkler experiment is a classic dynamic fracture experiment involving a pre-notched sample subjected 
+# to impact loading.
 
-# The geometrical parameters are defined as:
-# Sample length `l`, width `w`, and thickness `t`, with point spacing `Δx`, horizon size `$ δ $` and crack length `a`.
-l, w, t, Δx, δ, a = 100.0, 100 / 50, 50.0
+# ## Introduction
+#
+# The Kalthoff-Winkler experiment, conducted by John Kalthoff and Benedikt Winkler, is widely used to study fracture 
+# mechanics under high strain rates.
+# This setup provides valuable insights into the behavior of materials under dynamic loading conditions, making it a 
+# crucial experiment in the field of fracture mechanics.
+#
+# In this tutorial, we will simulate the Kalthoff-Winkler experiment using peridynamics, which is particularly suited 
+# for modeling problems involving discontinuities such as cracks.
 
-# Now a cuboid body with the specified dimensions is created using the bond-based material model with surface corrections:
+# ## Import Packages
+using Peridynamics
+using GLMakie # For visualization
+using LaTeXStrings # For LaTeX string formatting
+
+# ## Geometrical Parameters
+#
+# Define the sample length `l`, width `w`, and thickness `t`, with point spacing `Δx`, horizon size `$ δ $` and crack 
+# length `a`.
+l  = 200.0E-3  # Length of the sample (meters)
+w  = 100.0E-3  # Width of the sample (meters)
+t  =   9.0E-3  # Thickness of the sample (meters)
+Δx =   0.1E-3  # Discretization size (meters)
+δ  = 4.015Δx   # Horizon (meters)
+a  = 50.0E-3   # Crack length (meters)
+
+# ## Create the Body
+#
+# Create a cuboid body with the specified dimensions using the bond-based material model with surface corrections:
 pos, vol = uniform_box(l, w, t, Δx)
 body = Body(BBMaterial{EnergySurfaceCorrection}(), pos, vol)
 
-# Some plotting
-
+# ## Material Parameters
+#
 # The following material parameters are set:
 #
 # | material parameter | value |
@@ -29,32 +49,39 @@ body = Body(BBMaterial{EnergySurfaceCorrection}(), pos, vol)
 
 material!(body; horizon=4.015Δx, E=191.0e+9, rho=8000.0, Gc=22120.4)
 
-# Two point sets are defined to insert a crack between them at the left side of the domain (-x):
-point_set!(p -> -a / 2 - δ ≤ p[1] ≤ -a / 2 && 0 ≤ p[2] < wid / 2, body, :set_crack1_a)
-point_set!(p -> -a / 2 ≤ p[1] ≤ -a / 2 + δ && 0 ≤ p[2] < wid / 2, body, :set_crack1_b)
+# ## Define Pre-cracks
+#
+# Define the point sets to insert a crack at the left side of the domain (-x):
[…]
