### 2024-07-01 317602d
Draft: revival of `PointCloud`s
