+++ src/discretization/body.jl
@@ -63,14 +63,11 @@ Body{Material,PointParameters}
 - `single_dim_ics::Vector{SingleDimIC}`: Vector with defined initial conditions
 - `point_sets_precracks::Vector{PointSetsPreCrack}`: Vector with defined cracks
 """
-struct Body{M<:AbstractMaterial,P<:AbstractPointParameters} <: AbstractBody{M}
+struct Body{M<:AbstractMaterial,P<:AbstractPointParameters,PC} <: AbstractBody{M}
     name::Ref{Symbol}
     mat::M
-    n_points::Int
-    position::Matrix{Float64}
-    volume::Vector{Float64}
+    point_cloud::PC
     fail_permit::Vector{Bool}
-    point_sets::Dict{Symbol,Vector{Int}}
     point_params::Vector{P}
     params_map::Vector{Int}
     single_dim_bcs::Vector{SingleDimBC}
@@ -78,30 +75,28 @@ struct Body{M<:AbstractMaterial,P<:AbstractPointParameters} <: AbstractBody{M}
     single_dim_ics::Vector{SingleDimIC}
     point_sets_precracks::Vector{PointSetsPreCrack}
 
-    function Body(mat::M, position::AbstractMatrix, volume::AbstractVector) where {M}
+    function Body(mat::M, point_cloud::PC) where {M,PC}
         name = Symbol("")
-        n_points = length(volume)
-        check_pos_and_vol(n_points, position, volume)
-        fail_permit = fill(true, length(volume))
-        point_sets = Dict{Symbol,Vector{Int}}(:all_points => 1:length(volume))
+
+        fail_permit = fill(true, length(point_cloud.n_points))
 
         P = point_param_type(mat)
         point_params = Vector{P}()
-        params_map = zeros(Int, n_points)
+        params_map = zeros(Int, point_cloud.n_points)
 
         single_dim_bcs = Vector{SingleDimBC}()
         posdep_single_dim_bcs = Vector{PosDepSingleDimBC}()
         single_dim_ics = Vector{SingleDimIC}()
         point_sets_precracks = Vector{PointSetsPreCrack}()
 
-        new{M,P}(name, mat, n_points, position, volume, fail_permit, point_sets,
-                 point_params, params_map, single_dim_bcs, posdep_single_dim_bcs,
-                 single_dim_ics, point_sets_precracks)
+        new{M,P,PC}(name, mat, point_cloud, fail_permit, point_params, params_map,
+                    single_dim_bcs, posdep_single_dim_bcs, single_dim_ics,
+                    point_sets_precracks)
     end
 end
 
 function Base.show(io::IO, body::AbstractBody)
-    print(io, body.n_points, "-point Body{", material_type(body), "}")
+    print(io, n_points(body), "-point Body{", material_type(body), "}")
     if has_name(body)
         print(io, " with name `", get_name(body), "`")
     end
@@ -113,14 +108,14 @@ function Base.show(io::IO, ::MIME"text/plain", body::AbstractBody)
         show(io, body)
         return nothing
     end
-    print(io, body.n_points, "-point Body{", material_type(body), "}")
+    print(io, n_points(body), "-point Body{", material_type(body), "}")
     if has_name(body)
         print(io, " with name `", get_name(body), "`")
     end
     if has_point_sets(body) || has_params(body) || has_conditions(body)
         print(io, ":")
     end
-    for (name, points) in body.point_sets
+    for (name, points) in body.point_cloud.point_sets
         print(io, "\n  ", length(points), "-point set `", name, "`")
     end
     if has_params(body)
@@ -160,26 +155,6 @@ end
 
 @inline material_type(::AbstractBody{M}) where {M} = M
 
[…]
