+++ src/discretization/bond_system_corrections.jl
@@ -19,7 +19,6 @@ end
     return 1
 end
 
-
 #===========================================================================================
     ENERGY BASED SURFACE CORRECTION
 ===========================================================================================#
@@ -74,11 +73,11 @@ function calc_mfactor!(chunk::AbstractBodyChunk{BondSystem{EnergySurfaceCorrecti
     ka = 1.001
     for d in 1:3
         defposition = copy(system.position)
-        defposition[d,:] .*= ka
-        @views calc_stendens!(stendens[d,:], defposition, chunk)
+        defposition[d, :] .*= ka
+        @views calc_stendens!(stendens[d, :], defposition, chunk)
         for i in each_point_idx(chunk)
             params = get_params(chunk, i)
-            mfactor[d,i] = analytical_stendens(params) / stendens[d,i]
+            mfactor[d, i] = analytical_stendens(params) / stendens[d, i]
         end
     end
     return nothing
@@ -127,35 +126,27 @@ function calc_scfactor!(chunk::AbstractBodyChunk)
                 elseif abs(Δxijx) <= 1e-10
                     θ = 90 * π / 180
                 else
-                    θ = atan(abs(Δxijy)/abs(Δxijx))
+                    θ = atan(abs(Δxijy) / abs(Δxijx))
                 end
                 ϕ = 90 * π / 180
-                scx = (mfactor[1,i] + mfactor[1,j]) / 2
-                scy = (mfactor[2,i] + mfactor[2,j]) / 2
-                scz = (mfactor[3,i] + mfactor[3,j]) / 2
-                scr = sqrt(
-                    1/(
-                        (cos(θ) * sin(ϕ))^2 / scx^2 +
-                        (sin(θ) * sin(ϕ))^2 / scy^2 +
-                        cos(ϕ)^2 / scz^2
-                    )
-                )
+                scx = (mfactor[1, i] + mfactor[1, j]) / 2
+                scy = (mfactor[2, i] + mfactor[2, j]) / 2
+                scz = (mfactor[3, i] + mfactor[3, j]) / 2
+                scr = sqrt(1 / ((cos(θ) * sin(ϕ))^2 / scx^2 +
+                            (sin(θ) * sin(ϕ))^2 / scy^2 +
+                            cos(ϕ)^2 / scz^2))
             elseif abs(Δxijx) <= 1e-10 && abs(Δxijy) <= 1e-10
-                scz = (mfactor[3,i] + mfactor[3,j]) / 2
+                scz = (mfactor[3, i] + mfactor[3, j]) / 2
                 scr = scz
             else
-                θ = atan(abs(Δxijy)/abs(Δxijx))
-                ϕ = acos(abs(Δxijz)/L)
-                scx = (mfactor[1,i] + mfactor[1,j]) / 2
-                scy = (mfactor[2,i] + mfactor[2,j]) / 2
-                scz = (mfactor[3,i] + mfactor[3,j]) / 2
-                scr = sqrt(
-                    1/(
-                        (cos(θ) * sin(ϕ))^2 / scx^2 +
-                        (sin(θ) * sin(ϕ))^2 / scy^2 +
-                        cos(ϕ)^2 / scz^2
-                    )
-                )
+                θ = atan(abs(Δxijy) / abs(Δxijx))
+                ϕ = acos(abs(Δxijz) / L)
+                scx = (mfactor[1, i] + mfactor[1, j]) / 2
+                scy = (mfactor[2, i] + mfactor[2, j]) / 2
+                scz = (mfactor[3, i] + mfactor[3, j]) / 2
+                scr = sqrt(1 / ((cos(θ) * sin(ϕ))^2 / scx^2 +
[…]
