### 2024-07-01 1230594
Update `PointNeighbors.jl` to `v0.4`
