### 2024-07-01 434ddf9
Use `SemiParallelUpdate()` for contact search
