### 2024-07-01 12a4b09
Enhancements in logging bodies with multiple point parameters
