### 2024-07-01 6bb88d7
Full reworked vtk export path setup
