+++ test/auxiliary/test_io.jl
@@ -1,6 +1,10 @@
 @testitem "JobOptions" begin
     function tests_body_specific(body::Body, solver)
         default_fields_body = [field for field in Peridynamics.default_export_fields()]
+        body_name = string(Peridynamics.get_name(body))
+        vtk_root = joinpath("rootpath", "vtk")
+        vtk_filebase = joinpath(vtk_root,
+                                isempty(body_name) ? "timestep" : body_name * "_timestep")
 
         o = Dict{Symbol,Any}(:path => "rootpath", :freq => 10)
         jo = Peridynamics.get_job_options(body, solver, o)
@@ -10,6 +14,7 @@
         @test jo.logfile == joinpath("rootpath", "logfile.log")
         @test jo.freq == 10
         @test jo.fields == default_fields_body
+        @test jo.vtk_filebase == vtk_filebase
 
         o = Dict{Symbol,Any}(:path => "rootpath")
         jo = Peridynamics.get_job_options(body, solver, o)
@@ -19,6 +24,7 @@
         @test jo.logfile == joinpath("rootpath", "logfile.log")
         @test jo.freq == 10
         @test jo.fields == default_fields_body
+        @test jo.vtk_filebase == vtk_filebase
 
         o = Dict{Symbol,Any}(:freq => 10)
         msg = "if `freq` is specified, the keyword `path` is also needed!\n"
@@ -34,6 +40,7 @@
         @test jo.logfile == ""
         @test jo.freq == 0
         @test jo.fields == Vector{Symbol}()
+        @test jo.vtk_filebase == ""
 
         o = Dict{Symbol,Any}(:path => "rootpath", :freq => -10)
         msg = "`freq` should be larger than zero!\n"
@@ -49,6 +56,7 @@
         @test jo.logfile == joinpath("rootpath", "logfile.log")
         @test jo.freq == 10
         @test jo.fields == [:displacement, :b_ext]
+        @test jo.vtk_filebase == vtk_filebase
 
         o = Dict{Symbol,Any}(:path => "rootpath", :fields => :damage)
         jo = Peridynamics.get_job_options(body, solver, o)
@@ -58,6 +66,7 @@
         @test jo.logfile == joinpath("rootpath", "logfile.log")
         @test jo.freq == 10
         @test jo.fields == [:damage]
+        @test jo.vtk_filebase == vtk_filebase
 
         o = Dict{Symbol,Any}(:path => "rootpath", :fields => [:b_ext])
         jo = Peridynamics.get_job_options(body, solver, o)
@@ -67,6 +76,7 @@
         @test jo.logfile == joinpath("rootpath", "logfile.log")
         @test jo.freq == 10
         @test jo.fields == [:b_ext]
+        @test jo.vtk_filebase == vtk_filebase
 
         o = Dict{Symbol,Any}(:path => "rootpath", :fields => (:displacement, :b_ext, :abcd))
         @test_throws ArgumentError begin
@@ -79,8 +89,11 @@
     function tests_multibody_specific(ms::MultibodySetup, solver)
         default_fields_body = [field for field in Peridynamics.default_export_fields()]
         default_fields_multibody = Dict{Symbol,Vector{Symbol}}()
+        vtk_filebase = Dict{Symbol,String}()
+        vtk_root = joinpath("rootpath", "vtk")
         for body_name in Peridynamics.each_body_name(ms)
             default_fields_multibody[body_name] = default_fields_body
+            vtk_filebase[body_name] = joinpath(vtk_root, string(body_name) * "_timestep")
         end
 
[…]
