### 2024-07-02 b37b17d
Add more unit tests
