### 2024-07-02 f331c1f
Add version to logfiles
