### 2024-07-03 4a3f8e7
Major refactoring of boundary and initial conditions
