### 2024-07-03 d5ed25c
Refactoring conditions; closes #122
