+++ src/physics/boundary_conditions.jl
@@ -16,23 +16,6 @@ end
     return value
 end
 
-function override_eachother(a::SingleDimBC, b::SingleDimBC)
-    same_field = a.field === b.field
-    same_point_set = a.point_set === b.point_set
-    same_dim = a.dim == b.dim
-    return same_field && same_point_set && same_dim
-end
-
-function apply_bcs!(b::AbstractBodyChunk, time::Float64)
-    for bc in b.sdbcs
-        apply_bc!(b.storage, b.psets, bc, time)
-    end
-    for bc in b.pdsdbcs
-        apply_bc!(b.storage, b.psets, bc, b.system.position, time)
-    end
-    return nothing
-end
-
 function apply_bc!(s::AbstractStorage, psets::Dict{Symbol,Vector{Int}},
                    bc::SingleDimBC{F}, time::Float64) where {F}
     value = bc(time)
@@ -67,13 +50,6 @@ end
     return value
 end
 
-function override_eachother(a::PosDepSingleDimBC, b::PosDepSingleDimBC)
-    same_field = a.field === b.field
-    same_point_set = a.point_set === b.point_set
-    same_dim = a.dim == b.dim
-    return same_field && same_point_set && same_dim
-end
-
 function apply_bc!(s::AbstractStorage, psets::Dict{Symbol,Vector{Int}},
                    bc::PosDepSingleDimBC{F}, position::Matrix{Float64},
                    time::Float64) where {F}
@@ -93,6 +69,16 @@ end
     return nothing
 end
 
+function apply_boundary_conditions!(b::AbstractBodyChunk, time::Float64)
+    for bc in b.sdbcs
+        apply_bc!(b.storage, b.psets, bc, time)
+    end
+    for bc in b.pdsdbcs
+        apply_bc!(b.storage, b.psets, bc, b.system.position, time)
+    end
+    return nothing
+end
+
 function get_dim(dim::I) where {I<:Integer}
     in(dim, 1:3) || error("specified dimension should be 1=x, 2=y, or 3=z!\n")
     I <: UInt8 && return dim
@@ -106,19 +92,36 @@ function get_dim(dim::Symbol)
     return SYMBOL_TO_DIM[dim]
 end
 
-function add_condition!(conditions::Vector{B}, condition::B) where {B<:AbstractCondition}
-    if is_duplicate(condition, conditions)
-        error("duplicate conditions for point set $(condition.point_set)!\n")
-    end
+function add_boundary_condition!(body::AbstractBody, conditions::Vector{BC},
+                                 condition::BC) where {BC<:AbstractCondition}
+    check_boundary_condition_conflicts(body, condition)
     push!(conditions, condition)
     return nothing
 end
 
-function is_duplicate(condition::C, conditions::Vector{C}) where {C<:AbstractCondition}
-    for existing_condition in conditions
-        override_eachother(existing_condition, condition) && return true
+function check_boundary_condition_conflicts(body::AbstractBody, condition::BC) where {BC}
+    if has_boundary_condition_conflict(body, condition)
+        msg = "the specified condition conflicts with already existing conditions!\n"
+        throw(ArgumentError(msg))
+    end
+    return nothing
+end
+
+function has_boundary_condition_conflict(body::AbstractBody, condition::BC) where {BC}
+    for existing_condition in body.single_dim_bcs
+        conditions_conflict(body, existing_condition, condition) && return true
+    end
+    for existing_condition in body.posdep_single_dim_bcs
+        conditions_conflict(body, existing_condition, condition) && return true
     end
     return false
+end
+
[…]

[…]
