### 2024-07-03 c2b7821
Minor enhancements to error messages for wrong condition functions
