### 2024-07-03 2a2f88c
Improved the error message for wrong dimensions
