### 2024-07-03 664569f
Minor fixes to logs
