### 2024-07-03 36d8d66
Added tests on logs
