### 2024-07-04 c351570
Update bond_system.jl
An issue with the fracture criterion.
