+++ src/discretization/bond_system.jl
@@ -216,7 +216,7 @@ end
 @inline function stretch_based_failure!(storage::AbstractStorage, ::BondSystem,
                                         bond::Bond, params::AbstractPointParameters,
                                         ε::Float64, i::Int, bond_id::Int)
-    if ε > params.εc && bond.fail_permit
+    if abs(ε) > params.εc && bond.fail_permit
         storage.bond_active[bond_id] = false
     end
     storage.n_active_bonds[i] += storage.bond_active[bond_id]
