### 2024-07-04 abbd71f
Added a separate function for the influence function
