+++ README.md
@@ -45,6 +45,83 @@ To install `Peridynamics.jl`, follow these steps:
 
 We recommend looking at the [how-to guides](https://kaipartmann.github.io/Peridynamics.jl/stable/) and the [tutorials](https://kaipartmann.github.io/Peridynamics.jl/stable/) to start working with this package!
 
+## Tutorials Overview
+
+<style>
+.tutorial-grid {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
+    grid-gap: 20px;
+    padding: 25px;
+}
+
+.tutorial-element {
+    text-align: center;
+    /* background-color: #f9f9f9; */
+    background-color: rgb(240,240,240);;
+    padding: 20px;
+    border-radius: 10px;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+    transition: transform 0.3s, box-shadow 0.3s;
+}
+
+.tutorial-element:hover {
+    transform: scale(1.05);
+    box-shadow: 0 8px 16px rgba(0, 0, 0, 0.2);
+}
+
+.tutorial-element img {
+    width: 100%;
+    max-width: 300px;
+    height: auto;
+    border-radius: 10px;
+    margin-bottom: 10px;
+}
+
+.tutorial-element a {
+    color: #333;
+    text-decoration: none;
+    font-weight: bold;
+}
+
+.tutorial-element a:hover {
+    color: #007bff;
+}
+</style>
+
+<div class="tutorial-grid">
+
+<div class="tutorial-element">
+<a href="link_to_your_documentation#tutorial_tension_static">Tensile test quasi-static</a><br>
+<a href="link_to_your_documentation#tutorial_tension_static">
+    <img src="docs/src/assets/tension_static.gif" alt="Tensile test quasi-static">
+</a>
+</div>
+
+<div class="tutorial-element">
+<a href="link_to_your_documentation#tutorial_tension_dynfrac">Tensile test dynamic</a><br>
+<a href="link_to_your_documentation#tutorial_tension_dynfrac">
+    <img src="docs/src/assets/tension_dynfrac.gif" alt="Tensile test dynamic">
+</a>
+</div>
+
+<div class="tutorial-element">
+<a href="link_to_your_documentation#tutorial_tension_precrack">Tension with predefined crack</a><br>
+<a href="link_to_your_documentation#tutorial_tension_precrack">
+    <img src="docs/src/assets/tension_precrack_damage.gif" alt="Tension with predefined crack">
+</a>
+</div>
+
+<div class="tutorial-element">
+<a href="link_to_your_documentation#tutorial_logo">Peridynamics.jl logo</a><br>
+<a href="link_to_your_documentation#tutorial_logo">
+    <img src="docs/src/assets/logo.gif" alt="Peridynamics.jl logo">
+</a>
+</div>
+
+</div>
+
+
 ## Authors
 
 - <a href="https://orcid.org/0000-0002-5238-4355">Kai Partmann (University of Siegen) <img alt="ORCID logo" src="https://info.orcid.org/wp-content/uploads/2019/11/orcid_16x16.png" width="16" height="16" /></a>
