### 2024-07-04 6eb69bc
Delete links to old how-to guides
