### 2024-07-04 e664b32
Use the standard automatic docs
