+++ docs/src/api_reference.md
@@ -1,0 +1,13 @@
+# Peridynamics.jl API test
+
+```@meta
+CurrentModule = Peridynamics
+```
+
+```@index
+Pages = ["api_reference.md"]
+```
+
+```@autodocs
+Modules = [Peridynamics, Peridynamics.AbaqusMeshConverter, Peridynamics.VtkReader]
+```
