+++ docs/src/api_private.md
@@ -1,21 +1,0 @@
-# Private API
-
-Here, all non-exported types and methods are listed, which the average user does not need when using the package.
-
-```@meta
-
-```
-
-```@index
-Pages = ["api_private.md"]
-```
-
-## Types
-```@docs
-
-```
-
-## Functions
-```@docs
-
-```
