### 2024-07-04 070bd3a
added tutorial logo (closes #82)
