### 2024-07-04 017f017
Updated the docstrings of materials and abstract types
