+++ src/physics/bond_based.jl
@@ -1,35 +1,14 @@
 
 """
-    BBMaterial{Correction} <: AbstractMaterial
+    BBMaterial()
+    BBMaterial{Correction}()
 
-Material type for bond-based peridynamic simulations
+Construct the type `BBMaterial` used to specify the material of a [`Body`](@ref).
 
-# Type Parameters
-
-- `Correction`: Applied surface correction method
-    Available methods: - `NoCorrection`: No surface corrections are applied (default)
-                       - `EnergySurfaceCorrection`: Surface correction factors are applied
-                           based on the strain energy density
-
-# Allowed material parameters
-
-- `horizon::Float64`: Radius of point interactions
-- `rho::Float64`: Density
-- `E::Float64`: Young's modulus
-- `Gc::Float64`: Critical energy release rate
-- `epsilon_c::Float64`: Critical strain
-
-# Allowed export fields
-
-- `position::Matrix{Float64}`: Position of each point
-- `displacement::Matrix{Float64}`: Displacement of each point
-- `velocity::Matrix{Float64}`: Velocity of each point
-- `velocity_half::Matrix{Float64}`: Velocity parameter for Verlet time solver
-- `acceleration::Matrix{Float64}`: Acceleration of each point
-- `b_int::Matrix{Float64}`: Internal force density of each point
-- `b_ext::Matrix{Float64}`: External force density of each point
-- `damage::Vector{Float64}`: Damage of each point
-- `n_active_bonds::Vector{Int}`: Number of intact bonds for each point
+Possible correction methods are:
+- [`NoCorrection`](@ref): No correction is applied (default)
+- [`EnergySurfaceCorrection`](@ref): The energy based surface correction method of
+    Le and Bobaru (2018) is applied
 
 # Examples
 
@@ -39,6 +18,40 @@ BBMaterial{NoCorrection}()
 
 julia> mat = BBMaterial{EnergySurfaceCorrection}()
 BBMaterial{EnergySurfaceCorrection}()
+
+---
+
+```julia
+BBMaterial{Correction} <: AbstractBondSystemMaterial{Correction}
+```
+
+Material type for the bond-based peridynamics formulation.
+
+# Type Parameters
+
+- `Correction`: A correction algorithm of type [`AbstractCorrection`](@ref).
+
+# Allowed material parameters
+When using [`material!`](@ref) on a [`Body`](@ref) with `BBMaterial`, then the following
+parameters are allowed:
+- `horizon::Float64`: Radius of point interactions
+- `rho::Float64`: Density
+- `E::Float64`: Young's modulus
+- `Gc::Float64`: Critical energy release rate
+- `epsilon_c::Float64`: Critical strain
+
+# Allowed export fields
+When specifying the `fields` keyword of [`Job`](@ref) for a [`Body`](@ref) with
+`BBMaterial`, the following fields are allowed:
+- `position::Matrix{Float64}`: Position of each point
+- `displacement::Matrix{Float64}`: Displacement of each point
+- `velocity::Matrix{Float64}`: Velocity of each point
+- `velocity_half::Matrix{Float64}`: Velocity parameter for Verlet time solver
+- `acceleration::Matrix{Float64}`: Acceleration of each point
+- `b_int::Matrix{Float64}`: Internal force density of each point
+- `b_ext::Matrix{Float64}`: External force density of each point
+- `damage::Vector{Float64}`: Damage of each point
+- `n_active_bonds::Vector{Int}`: Number of intact bonds of each point
 ```
 
 """
