+++ src/Peridynamics.jl
@@ -48,54 +48,74 @@ function __init__()
 end
 
 """
-TODO
+    AbstractMaterial
+
+Base for types which define materials.
 """
 abstract type AbstractMaterial end
 
 """
-TODO
+    AbstractSpatialSetup
+
+Base for types which define the spatial setup.
 """
 abstract type AbstractSpatialSetup end
 
 """
-TODO
+    AbstractBody{T<:AbstractMaterial} <: AbstractSpatialSetup
+
+Base for types which define a body.
 """
 abstract type AbstractBody{T<:AbstractMaterial} <: AbstractSpatialSetup end
 
 """
-TODO
+    AbstractMultibodySetup <: AbstractSpatialSetup
+
+Base for types which define a multibody setup.
 """
 abstract type AbstractMultibodySetup <: AbstractSpatialSetup end
 
 """
-TODO
+    AbstractParameterSetup
+
+Base for types which define a parameter setup.
 """
 abstract type AbstractParameterSetup end
 
 """
-TODO
+    AbstractPointParameters <: AbstractParameterSetup
+
+Base for types which define point parameters for specific materials.
 """
 abstract type AbstractPointParameters <: AbstractParameterSetup end
 abstract type AbstractParamSpec end
 
 """
-TODO
+    AbstractTimeSolver
+
+Base for types which define a time solver.
 """
 abstract type AbstractTimeSolver end
 
 """
-TODO
+    AbstractJob
+
+Base for types which define a job.
 """
 abstract type AbstractJob end
 abstract type AbstractJobOptions end
 
 """
-TODO
+    AbstractSystem
+
+Base for types which define a system.
 """
 abstract type AbstractSystem end
 
 """
-TODO
+    AbstractPredefinedCrack
+
+Base for types which define a predefined crack.
 """
 abstract type AbstractPredefinedCrack end
 abstract type AbstractBodyChunk{S<:AbstractSystem,T<:AbstractMaterial} end
@@ -110,24 +130,39 @@ abstract type AbstractMPIBodyDataHandler{Sys,M,P,S} <: AbstractMPIDataHandler en
 abstract type AbstractMPIMultibodyDataHandler <: AbstractMPIDataHandler end
 
 """
-TODO
+    AbstractCorrection
+
+Base for types which define a correction algorithm suitable for materials that are a
+subtype of [`AbstractBondSystemMaterial`](@ref) and use the [`BondSystem`](@ref).
 """
 abstract type AbstractCorrection end
 
 """
-TODO
+    AbstractStorage
+
+Base for types which define a storage.
 """
 abstract type AbstractStorage end
 
 """
-TODO
+    AbstractStorage
+
+Base for types which define a condition.
 """
 abstract type AbstractCondition end
 
 """
-TODO
+    AbstractBondSystemMaterial{Correction} <: AbstractMaterial
+
+Base for types which define a material that uses the [`BondSystem`](@ref).
 """
 abstract type AbstractBondSystemMaterial{Correction} <: AbstractMaterial end
+
+"""
+    AbstractInteractionSystemMaterial <: AbstractMaterial
+
+Base for types which define a material that uses the [`InteractionSystem`](@ref).
+"""
 abstract type AbstractInteractionSystemMaterial <: AbstractMaterial end
 
 include("auxiliary/function_arguments.jl")
