### 2024-07-04 f1be4c5
Does not print private types in show methods; added docs
