+++ src/discretization/body.jl
@@ -2,38 +2,27 @@
     Body(material, position, volume)
     Body(material, inp_file)
 
-Creates a body for use in peridynamic calculation
+Constructs a `Body` for a peridynamics simulation.
 
 # Arguments
-
-- `material::AbstractMaterial`: Specifies which material model is used
-- `position::AbstractMatrix`: 3×n matrix with position of each point
-- `volume::AbstractVector`: Vector with volume of each point
-- `inp_file::AbstractString`: A Abaqus input file containing meshes, imported with
-    [`read_inp`](@ref)
+- `material::AbstractMaterial`: The material which is defined for the whole body.
+- `position::AbstractMatrix`: A `3×n` matrix with the point position of the `n` points.
+- `volume::AbstractVector`: A vector with the volume of each point.
+- `inp_file::AbstractString`: An Abaqus input file containing meshes, imported with
+    [`read_inp`](@ref).
 
 # Throws
-
-- Error if n_points = 0
-- `DimensionMismatch`: Error if dimension of position != 3
+- Errors if the number of points is not larger than zero
+- Errors if `position` is not a `3×n` matrix and has the same length as `volume`
+- Errors if `position` or `volume` contain `NaN` values
 
 # Example
 
 ```julia-repl
-julia> l, Δx, = 1.0, 1/50;
-
-julia> pos, vol = uniform_box(l, l, 0.1l, Δx);
-
-julia> b = Body(BBMaterial(), pos, vol);
-
-julia> b
-Body{BBMaterial, Peridynamics.BBPointParameters}(BBMaterial(), 12500,
-[-0.49 -0.47 … 0.47 0.49; -0.49 -0.49 … 0.49 0.49; -0.04 -0.04 … 0.04 0.04],
-[8.000000000000001e-6, 8.000000000000001e-6,  …  8.000000000000001e-6],
-Bool[1, 1, 1, 1, 1, 1, 1, 1, 1, 1  …  1, 1, 1, 1, 1, 1, 1, 1, 1, 1],
-Dict{Symbol, Vector{Int64}}(), Peridynamics.BBPointParameters[],
-[0, 0, 0, 0, 0, 0, 0, 0, 0, 0  …  0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-Peridynamics.SingleDimBC[], Peridynamics.SingleDimIC[], Peridynamics.PointSetsPreCrack[])
+julia> Body(BBMaterial(), rand(3, 10), rand(10))
+10-point Body{BBMaterial{NoCorrection}}:
+  1 point set(s):
+    10-point set `all_points`
 ```
 
 ---
@@ -54,17 +43,24 @@ Body{Material,PointParameters}
 
 # Fields
 
-- `mat::Material`: Specified material model
-- `n_points::Int`: Number of material points that represent the body
-- `position::Matrix{Float64}`: 3×n_points matrix with position for each point
-- `volume::Vector{Float64}`: Vector with volume for each point
-- `fail_permit::Vector{Bool}`: Vector that describes if failure is allowed for each point
-- `point_sets::Dict{Symbol,Vector{Int}}`: Dictionary containing the defined point sets
-- `point_params::Vector{PointParameters}`: Vector with material parameter sets
-- `params_map::Vector{Int}`: Vector that assigns a material parameter set to each point
-- `single_dim_bcs::Vector{SingleDimBC}`: Vector with defined boundary conditions
-- `single_dim_ics::Vector{SingleDimIC}`: Vector with defined initial conditions
-- `point_sets_precracks::Vector{PointSetsPreCrack}`: Vector with defined cracks
+- `mat::Material`: The material formulation.
+- `n_points::Int`: The number of points that in the body.
+- `position::Matrix{Float64}`: A `3×n_points` matrix with the point position of the points.
[…]
