### 2024-07-04 d9fcf2e
Removed empty docstrings from internal functions
