### 2024-07-05 80abaf0
BibTeX cleanup
