+++ docs/src/refs.bib
@@ -6,11 +6,9 @@
     pages = {175-209},
     year = {2000},
     issn = {0022-5096},
-    doi = {https://doi.org/10.1016/S0022-5096(99)00029-0},
-    url = {https://www.sciencedirect.com/science/article/pii/S0022509699000290},
+    doi = {10.1016/S0022-5096(99)00029-0},
     author = {S.A. Silling},
     keywords = {Dynamics, Elastic material, Fracture, Stability and bifurcation, Stress waves},
-    abstract = {Some materials may naturally form discontinuities such as cracks as a result of deformation. As an aid to the modeling of such materials, a new framework for the basic equations of continuum mechanics, called the ‘peridynamic’ formulation, is proposed. The propagation of linear stress waves in the new theory is discussed, and wave dispersion relations are derived. Material stability and its connection with wave propagation is investigated. It is demonstrated by an example that the reformulated approach permits the solution of fracture problems using the same equations either on or off the crack surface or crack tip. This is an advantage for modeling problems in which the location of a crack is not known in advance.}
 }
 
 @article{Silling2005,
@@ -33,11 +31,9 @@
     year = {2005},
     note = {Special Issue in Honour of C.O. Horgan},
     issn = {0020-7462},
-    doi = {https://doi.org/10.1016/j.ijnonlinmec.2004.08.004},
-    url = {https://www.sciencedirect.com/science/article/pii/S0020746204001027},
+    doi = {10.1016/j.ijnonlinmec.2004.08.004},
     author = {S.A. Silling and F. Bobaru},
     keywords = {Elasticity, Membranes, Dynamic fracture, Rubber},
-    abstract = {The peridynamic theory of continuum mechanics allows damage, fracture, and long-range forces to be treated as natural components of the deformation of a material. In this paper, the peridynamic approach is applied to small thickness two- and one-dimensional structures. For membranes, a constitutive model is described appropriate for rubbery sheets that can form cracks. This model is used to perform numerical simulations of the stretching and dynamic tearing of membranes. A similar approach is applied to one-dimensional string like structures that undergrow stretching, bending, and failure. Long-range forces similar to van der Waals interactions at the nanoscale influence the equilibrium configurations of these structures, how they deform, and possibly self-assembly.}
 }
 
 @article{Silling2007,
