### 2024-07-05 79153f7
Fail on missing docstrings etc.
