+++ docs/make.jl
@@ -86,7 +86,7 @@ makedocs(;
     # ONLY DURING EARLY DEVELOPMENT:
     # MISSING DOCSTRINGS DUE TO API CHANGES DO NOT RESULT IN ERRORS
     # THAT STOP THE DOCUMENTATION BUILD PROCESS!
-    warnonly = true,
+    # warnonly = true,
 )
 
 if !LIVE_MODE
