### 2024-07-05 a35c54a
Move bibliography to each individual page
