+++ docs/src/nosbased.md
@@ -23,3 +23,8 @@ Using the calculated variables, the force vector state can now be determined by 
 \boldsymbol{t}^i = \omega \boldsymbol{P}^i  \left(\boldsymbol{K}^i\right)^{-1} \boldsymbol{\Delta X}^{ij} \; .
 ```
 
+## References
+```@bibliography
+Pages = ["nosbased.md"]
+Canonical = false
+```
