### 2024-07-05 97b64bf
Updated all docstrings; closes #60; closes #76; closes #92
