+++ src/core/submit.jl
@@ -6,27 +6,11 @@ const SUBMIT_KWARGS = (:quiet,)
 Submits the job to start calculations
 
 # Arguments
-
-- `job::Job`: Job that contains all defined parameters and conditions
+- `job::Job`: Job that contains all defined parameters and conditions.
 
 # Keywords
 
-- `quiet::Bool`: If `:quiet=true`, no outputs are printed in the terminal.
-
-# Throws
-
-- Error if keyword is not allowed
-
-# Example
-
-```julia-repl
-julia> job = Job(b, vv;
-           path=joinpath(@__DIR__, "results", "mode_I"),
-           fields=(:displacement, :damage));
-
-julia> submit(job);
-solve... 100%|████████████████████| Time: 0:00:11
-```
+- `quiet::Bool`: If `true`, no outputs are printed in the terminal.
 """
 function submit(job::Job; kwargs...)
     o = Dict{Symbol,Any}(kwargs)
