+++ src/VtkReader/VtkReader.jl
@@ -288,14 +288,12 @@ Read vtu or pvtu file containing simulation results of a time step.
 # Examples
 
 ```julia-repl
-julia> read_vtk("ExampleSimulation_t3000.vtu")
-Dict{String, VecOrMat{Float64}} with 6 entries:
-  "position"     => [-0.497302 -0.497303 … 0.497303 0.497302; -0.0225023 -0…
-  "damage"       => [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0  …  0…
-  "time"         => [0.000237879]
-  "displacement" => [0.00019766 0.000196793 … -0.000196793 -0.00019766; -2.…
-  "velocity"     => [-2.55436 -1.5897 … 1.5897 2.55436; 0.827107 0.234996 ……
-  "b_int"        => [-7.00131e9 -8.45411e9 … 8.45411e9 7.00131e9; -5.0164e8…
+julia> read_vtk("results/fragmenting_cylinder/vtk/timestep_000520.pvtu")
+Dict{Symbol, VecOrMat{Float64}} with 4 entries:
+  :position     => [0.0263309 0.027315 … 0.0293543 0.030339; 0.000292969 0.000294475…
+  :displacement => [0.00583334 0.00581883 … 0.00585909 0.00584271; -0.000162852 -0.0…
+  :damage       => [0.616071, 0.569343, 0.528571, 0.463415, 0.438776, 0.553571, 0.56…
+  :time         => [9.69363e-5]
 ```
 """
 function read_vtk(file::AbstractString)
