+++ README.md
@@ -1,7 +1,7 @@
 <img src="docs/src/assets/logo.png" width="450" />
 
 # Peridynamics.jl
-A high-level Julia package for multithreaded peridynamics simulations
+A high-level Julia package for parallel peridynamics simulations
 
 **Documentation:**\
 [![Stable](https://img.shields.io/badge/docs-stable-blue.svg)](https://kaipartmann.github.io/Peridynamics.jl/stable/)
@@ -18,11 +18,12 @@ A high-level Julia package for multithreaded peridynamics simulations
 **Citation:**\
 [![DOI](https://zenodo.org/badge/503281781.svg)](https://zenodo.org/badge/latestdoi/503281781)
 
-## What is peridynamics?
-Peridynamics is a non-local formulation of continuum mechanics where material points represent the continuum, and the relative displacements and forces are governed by an integro-differential equation that allows discontinuities. As a result, peridynamics is particularly well-suited for dynamic fracture simulations involving numerous cracks.
-
-## Who can benefit from this package?
-This package is designed for anyone interested in performing peridynamics simulations. While the current feature set provides a solid foundation, we are continuously working to enhance and expand the capabilities of `Peridynamics.jl`. We encourage you to open an issue or submit a pull request to share your feedback or contribute to making this package even more valuable to the community!
+## Main Features
+- 🎯 Dynamic and quasi-static analysis
+- 🪨 Multiple peridynamics formulations and material models
+- 🎳 Multibody contact simulations
+- 🤓 User friendly API that captures many errors before submitting simulations
+- 🚀 Enhanced HPC capabilities with MPI and multiple threads
 
 ## Installation
 
@@ -41,9 +42,30 @@ To install `Peridynamics.jl`, follow these steps:
 
 5. Press `Backspace` or `Ctrl + C` to exit the package manager.
 
-## Getting Started
+## Tutorials
 
-We recommend looking at the [how-to guides](https://kaipartmann.github.io/Peridynamics.jl/stable/) and the [tutorials](https://kaipartmann.github.io/Peridynamics.jl/stable/) to start working with this package!
+<table align="center" border="0">
+  <tr>
+  </tr>
+  <tr>
+    <td align="center">
+      <img src="docs/src/assets/tension_static.gif" style="width: 80% !important;"/><br><figcaption>Tensile test quasi-static</figcaption>
+    </td>
+    <td align="center">
+      <img src="docs/src/assets/tension_dynfrac.gif" style="width: 80% !important;"/><br><figcaption>Tensile test dynamic</figcaption>
+    </td>
+  </tr>
+  <tr>
+  </tr>
+  <tr>
+    <td align="center">
+      <img src="docs/src/assets/tension_precrack_damage.gif" style="width: 80% !important;"/><br><figcaption>tension with predefined crack</figcaption>
+    </td>
+    <td align="center">
+      <img src="docs/src/assets/logo.gif" style="width: 80% !important;"/><br><figcaption>Periydnamics.jl logo</figcaption>
+    </td>
+  </tr>
+</table>
 
 ## Authors
 
@@ -51,7 +73,7 @@ We recommend looking at the [how-to guides](https://kaipartmann.github.io/Peridy
 - <a href="https://orcid.org/0009-0004-9195-0112">Manuel Dienst (University of Siegen) <img alt="ORCID logo" src="https://info.orcid.org/wp-content/uploads/2019/11/orcid_16x16.png" width="16" height="16" /></a>
[…]
