### 2024-07-05 091473e
Replaced all images in explanation section of docs
