+++ docs/src/references.bib
@@ -57,3 +57,219 @@
 	journal = {Journal of the Mechanics and Physics of Solids},
 	doi = {10.1016/j.jmps.2019.06.016}
 }
+
+@article{Bezanson2017julia,
+  title     = {Julia: A fresh approach to numerical computing},
+  author    = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
+  doi       = {10.1137/141000671},
+  journal   = {SIAM review},
+  volume    = {59},
+  number    = {1},
+  pages     = {65--98},
+  year      = {2017},
+  publisher = {SIAM}
+}
+
+@article{Diehl2019,
+  author   = {Diehl, Patrick and Prudhomme, Serge and L{\'e}vesque, Martin},
+  doi      = {10.1007/s42102-018-0004-x},
+  journal  = {Journal of Peridynamics and Nonlocal Modeling},
+  language = {en},
+  number   = {1},
+  pages    = {14--35},
+  title    = {A {Review} of {Benchmark} {Experiments} for the {Validation} of {Peridynamics} {Models}},
+  volume   = {1},
+  year     = {2019}
+}
+
+@article{Friebertshaeuser2022AIMS,
+  title   = {Dynamic fracture with continuum-kinematics-based peridynamics},
+  journal = {AIMS Materials Science},
+  volume  = {9},
+  number  = {6},
+  pages   = {791-807},
+  year    = {2022},
+  doi     = {10.3934/matersci.2022049},
+  author  = {Kai {Friebertsh{\"a}user (Partmann)} and Christian Wieners and Kerstin Weinberg}
+}
+
+@article{Friebertshaeuser2022PAMM,
+  title     = {Dynamic fracture with a continuum-kinematics-based peridynamic and a phase-field approach},
+  author    = {{Friebertsh{\"a}user (Partmann)}, Kai and Thomas, Marita and Tornquist, Sven and Weinberg, Kerstin and Wieners, Christian},
+  journal   = {PAMM},
+  volume    = {22},
+  number    = {1},
+  pages     = {e202200217},
+  year      = {2023},
+  publisher = {Wiley Online Library},
+  doi       = {10.1002/pamm.202200217}
+}
+
+@article{Javili2019,
+  author  = {Javili, Ali and McBride, A. and Steinmann, Paul},
+  year    = {2019},
+  month   = {07},
+  pages   = {},
+  title   = {Continuum-kinematics-inspired peridynamics. Mechanical problems},
+  volume  = {131},
+  journal = {Journal of the Mechanics and Physics of Solids},
+  doi     = {10.1016/j.jmps.2019.06.016}
+}
+
+@article{Javili2019Review,
+  author   = {Javili, Ali and Morasata, Rico and Oterkus, Erkan and Oterkus, Selda},
+  doi      = {10.1177/1081286518803411},
+  journal  = {Mathematics and Mechanics of Solids},
+  language = {en},
+  number   = {11},
+  pages    = {3714--3739},
+  title    = {Peridynamics review},
+  volume   = {24},
+  year     = {2019}
+}
+
+@article{Jha2021,
+  doi       = {10.21105/joss.03020},
+  year      = {2021},
+  publisher = {The Open Journal},
+  volume    = {6},
+  number    = {65},
+  pages     = {3020},
+  author    = {Prashant K. Jha and Patrick Diehl},
+  title     = {NLMech: Implementation of finite difference/meshfree discretization of nonlocal fracture models},
+  journal   = {Journal of Open Source Software}
+}
+
+@article{Kilic2010,
+  title   = {An adaptive dynamic relaxation method for quasi-static simulations using the peridynamic theory},
[…]
