+++ docs/src/expl_general_pd.md
@@ -1,28 +1,38 @@
 # [Basics of peridynamics theory](@id expl_pd_basics)
 
-Peridynamics is a nonlocal formulation of continuum mechanics whereby in the discrete version a body is represented by a group of interconnected material points.
-The basic idea is that each point interacts with any other point within the distance of the horizon $\delta$. 
-These interactions called *bonds* provoke forces between points which can be calculated and result in internal forces.
+Peridynamics is a nonlocal continuum mechanics formulation, which was introduced by Silling [Silling2000](@cite).
+It has gained increased popularity as an approach for modeling fracture.
+The deformation of the solid is described by integro-differential equations that are also fulfilled for discontinuities, making it very capable of modeling crack propagation and fragmentation with large displacements.
+Much peridynamics research has been done in recent years, summarized in various review papers and books [Diehl2019,Javili2019Review,Madenci2014](@cite).
 
 ```@raw html
-<img src="../assets/PDBody.png" width="400"/>
+<img src="https://github.com/kaipartmann/Peridynamics.jl/assets/68582683/728da1f0-4750-4ab6-a430-9b206e475577" width="300"/>
 ```
 
-Each point $i$ of the body $\mathcal{B}_0$ in initial configuration is characterized by the vector $\boldsymbol{X}^{i}$. The neighborhood $\mathcal{H}^i$ of $i$ contains all points $j$ that are within the distance of horizon $\delta$ to $i$:
-
+Typically, in peridynamics the continuum is discretized by material points.
+Points interact only with other points inside of their specified *neighborhood* or *point family* $\mathcal{H}$, which is defined as the set of points inside a sphere with the radius $\delta$, also named the *horizon*.
+The interaction of the point $\boldsymbol{X}$ with its neighbor $\boldsymbol{X}'$ is called *bond* and defined as
 ```math
-    \mathcal{H}^i = \{ \boldsymbol{X}^j \in \mathcal{B}_0 \; | \; 0 < \left| \boldsymbol{X}^j - \boldsymbol{X}^i \right| \leq \delta \} \quad \forall \; \boldsymbol{X}^i \in \mathcal{B}_0 \; .
+\boldsymbol{\Delta X} = \boldsymbol{X}' - \boldsymbol{X} \; .
 ```
-
-All of those combinations of points $i$ and $j$ form bonds, characterized by 
-$\boldsymbol{\Delta X}^{ij}= \boldsymbol{X}^j - \boldsymbol{X}^i$.
-
-In current configuration $\mathcal{B}_t$, lowercase letters are used to describe points and bonds: $\boldsymbol{x} ^ {i} $ and $\boldsymbol{\Delta x}^{ij}$.
-
-The peridynamic equation of motion is established for each material point $i$ and equals [Silling2000](@cite)
+The equation of motion reads
 ```math
-\rho \ddot{\boldsymbol{u}} (\boldsymbol{X}^{i} , t) = \boldsymbol{b}^{\mathrm{int}} (\boldsymbol{X}^{i} , t) + \boldsymbol{b}^{\mathrm{ext}} (\boldsymbol{X}^{i} , t) 
-    \qquad \forall \; \boldsymbol{X} ^ {i} \in \mathcal{B}_0 \; , \; t \geq 0 \; .
+\varrho \, \boldsymbol{\ddot{u}}(\boldsymbol{X},t) = \boldsymbol{b}^{\mathrm{int}}(\boldsymbol{X},t) + \boldsymbol{b}^{\mathrm{ext}}(\boldsymbol{X},t) \; ,
 ```
+with the mass density $\varrho$, the point acceleration vector $\boldsymbol{\ddot{u}}$, and the point force density vectors $\boldsymbol{b}^{\mathrm{int}}$ and $\boldsymbol{b}^{\mathrm{ext}}$.
[…]
