### 2024-07-05 ce5065d
Updated the image paths in ParaView How-to
