### 2024-07-06 8852c45
Updated the tensile test tutorials
