### 2024-07-06 27fc3c7
Update tutorial table
