+++ dev/references.bib
@@ -1,0 +1,275 @@
+@article{Silling2000,
+    title = {Reformulation of elasticity theory for discontinuities and long-range forces},
+    journal = {Journal of the Mechanics and Physics of Solids},
+    volume = {48},
+    number = {1},
+    pages = {175-209},
+    year = {2000},
+    issn = {0022-5096},
+    doi = {10.1016/S0022-5096(99)00029-0},
+    author = {S.A. Silling},
+    keywords = {Dynamics, Elastic material, Fracture, Stability and bifurcation, Stress waves},
+}
+
+@article{Silling2005,
+    author = {Silling, S.A. and Askari, E.},
+    year = {2005},
+    month = {06},
+    pages = {1526-1535},
+    title = {A meshfree method based on the peridynamic model of solid mechanics},
+    volume = {83},
+    journal = {Computers \& Structures},
+    doi = {10.1016/j.compstruc.2004.11.026}
+}
+
+@article{Silling2005a,
+    title = {Peridynamic modeling of membranes and fibers},
+    journal = {International Journal of Non-Linear Mechanics},
+    volume = {40},
+    number = {2},
+    pages = {395-409},
+    year = {2005},
+    note = {Special Issue in Honour of C.O. Horgan},
+    issn = {0020-7462},
+    doi = {10.1016/j.ijnonlinmec.2004.08.004},
+    author = {S.A. Silling and F. Bobaru},
+    keywords = {Elasticity, Membranes, Dynamic fracture, Rubber},
+}
+
+@article{Silling2007,
+    author = {Silling, S. and Epton, Michael and Weckner, Olaf and Xu, Jifeng and Askari, E.},
+    year = {2007},
+    month = {01},
+    pages = {151-184},
+    title = {Peridynamic States and Constitutive Modeling},
+    volume = {88},
+    journal = {Journal of Elasticity},
+    doi = {10.1007/s10659-007-9125-1}
+}
+
+@article{Javili2019,
+	author = {Javili, Ali and McBride, A. and Steinmann, Paul},
+	year = {2019},
+	month = {07},
+	pages = {},
+	title = {Continuum-kinematics-inspired peridynamics. Mechanical problems},
+	volume = {131},
+	journal = {Journal of the Mechanics and Physics of Solids},
+	doi = {10.1016/j.jmps.2019.06.016}
+}
+
+@article{Bezanson2017julia,
+  title     = {Julia: A fresh approach to numerical computing},
+  author    = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
+  doi       = {10.1137/141000671},
+  journal   = {SIAM review},
+  volume    = {59},
+  number    = {1},
+  pages     = {65--98},
+  year      = {2017},
+  publisher = {SIAM}
+}
+
+@article{Diehl2019,
+  author   = {Diehl, Patrick and Prudhomme, Serge and L{\'e}vesque, Martin},
+  doi      = {10.1007/s42102-018-0004-x},
+  journal  = {Journal of Peridynamics and Nonlocal Modeling},
+  language = {en},
+  number   = {1},
+  pages    = {14--35},
+  title    = {A {Review} of {Benchmark} {Experiments} for the {Validation} of {Peridynamics} {Models}},
+  volume   = {1},
+  year     = {2019}
+}
+
+@article{Friebertshaeuser2022AIMS,
+  title   = {Dynamic fracture with continuum-kinematics-based peridynamics},
+  journal = {AIMS Materials Science},
+  volume  = {9},
+  number  = {6},
+  pages   = {791-807},
+  year    = {2022},
+  doi     = {10.3934/matersci.2022049},
[…]
