### 2024-07-07 009538a
Improved the Peridynamics.jl logo
