### 2024-07-08 b4a7233
Added log for each system when using a multibody analysis
