### 2024-07-08 b21de58
Add `@nospecialize` to each `Base.show` method
