### 2024-07-08 3d2a00b
Add body name to system logs
