### 2024-07-08 9df0313
Minor docstring changes for clarity
