### 2024-07-09 7acaf51
Shorten the installation section
