### 2024-07-09 987dbdf
Enhanced docstrings
