### 2024-07-09 fb99d4c
Corrections to tutorial_kalthoff-winkler_dynfrac.jl per kaipartmann direction
Addressed comments per Kai's direction.
