### 2024-07-09 94b0a3f
Formatting, small parameter changes and visualization
