+++ docs/src/literate/tutorial_kalthoff-winkler_dynfrac.jl
@@ -1,35 +1,38 @@
 # # [Kalthoff-Winkler experiment](@id tutorial_kalthoff-winkler)
 #
-# This tutorial demonstrates how to set up and run the Kalthoff-Winkler experiment using the Peridynamics.jl package.
-# The Kalthoff-Winkler experiment is a classic dynamic fracture experiment involving a pre-notched sample subjected 
-# to impact loading.
+# This tutorial demonstrates how to set up and run the Kalthoff-Winkler experiment using the
+# Peridynamics.jl package. The Kalthoff-Winkler experiment is a classic dynamic fracture
+# experiment involving a pre-notched sample subjected to impact loading.
 
 # ## Introduction
 #
-# The Kalthoff-Winkler experiment is widely used to study fracture mechanics under high strain rates.
-# This setup provides valuable insights into the behavior of materials under dynamic loading conditions, making it an
-# interesting experiment in the field of fracture mechanics.
+# The Kalthoff-Winkler experiment is widely used to study fracture mechanics under high
+# strain rates. This setup provides valuable insights into the behavior of materials under
+# dynamic loading conditions, making it an interesting experiment in the field of fracture
+# mechanics.
 #
-# In this tutorial, we will simulate the Kalthoff-Winkler experiment using peridynamics without the impactor. To do so,
-# first we would import the packages.
+# In this tutorial, we will simulate the Kalthoff-Winkler experiment using peridynamics
+# without the impactor. To do so, first we would import the packages.
 
 # ## Import Packages
 using Peridynamics
 
 # ## Geometrical Parameters
 #
-# Define the sample length `l`, width `w`, and thickness `t`, with point spacing `Δx`, horizon size `$ δ $` and crack 
+# Define the sample length `l`, width `w`, and thickness `t`, with point spacing `Δx`,
+# horizon size `$ δ $` and crack
 # length `a`.
 l  = 200.0E-3  # Length of the sample (meters)
 w  = 100.0E-3  # Width of the sample (meters)
 t  =   9.0E-3  # Thickness of the sample (meters)
 Δx =   1.0E-3  # Discretization size (meters)
-δ  = 4.015Δx   # Horizon (meters)
-a  = 50.0E-3   # Crack length (meters)
+δ  =  4.015Δx  # Horizon (meters)
+a  =  50.0E-3  # Crack length (meters)
 
 # ## Create the Body
 #
-# Create a body with the specified dimensions using the bond-based material model with surface corrections:
+# Create a body with the specified dimensions using the bond-based material model with
+# surface corrections:
 pos, vol = uniform_box(l, w, t, Δx)
 body = Body(BBMaterial{EnergySurfaceCorrection}(), pos, vol)
 
@@ -42,9 +45,8 @@ body = Body(BBMaterial{EnergySurfaceCorrection}(), pos, vol)
 # | Horizon $ δ $ | $4.015 \cdot Δx$ |
 # | Young's modulus $E$ | $ 191\cdot 10^{9} \, \mathrm{Pa}$ |
 # | Density $ρ$ | $ 8000 ,\mathrm{kg}\,\mathrm{m}^{-3}$ |
-# | Griffith's parameter $G_c$ | $22,170 \, \mathrm{J} \, \mathrm{m}^{-2}$ |
-
-material!(body; horizon=4.015Δx, E=191.0e+9, rho=8000.0, Gc=22120.4)
+# | Critical stretch $\varepsilon_c$ | $0.01$ |
+material!(body; horizon=δ, E=191.0e9, rho=8000.0, epsilon_c=0.015)
 
 # ## Define Pre-cracks
 #
[…]
