### 2024-07-09 5019049
Tutorials overview style fix
