### 2024-07-09 95b1319
Add new line in Kalthoff-Winkler tutorial cell
