### 2024-07-09 dba52cf
Fixed and cleaned some equations
