+++ docs/src/expl_bondbased.md
@@ -5,7 +5,7 @@ The initial version of peridynamics is the bond-based (BB) formulation. [Silling
 Here, a pairwise force function $\boldsymbol{f}$ is defined and calculated for each bond of two material points, which depends on the strain of the bond and is aligned in its direction: 
 
 ```math
- \boldsymbol{f} = d^{ij} \cdot c \cdot \varepsilon^{ij} \cdot \boldsymbol{n} \; .
+ \boldsymbol{f} = d^{ij} \, c \, \varepsilon^{ij} \, \boldsymbol{n} \; .
 ```
 
 with
@@ -23,7 +23,7 @@ In this expression, bond failure $d^{ij}$ represents whether the bond between po
 or damaged ($d^{ij}=0$). The direction vector $\boldsymbol{n}$ is oriented in the direction of the bond. 
 Furthermore, the micro-modulus constant [Silling2005](@cite)
 ```math
-c = \frac{18 \cdot \kappa}{\pi \cdot \delta^4} \;
+c = \frac{18 \, \kappa}{\pi \, \delta^4} \;
 ```
 and the strain of the bond [Silling2005a](@cite)
 ```math
