### 2024-07-09 204e54a
Minor code cleanup
