### 2024-07-09 cc4e246
Update version to `v0.3.0`
