### 2024-07-10 69b1345
Fixes #151 also for `vtu` files
