### 2024-07-16 #150 @luraess
@HaoZeke, @sajavicius 👋 Would you be interested in and available to review this JuliaCon Proceeding submission? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://juliacon.github.io/proceedings-guide/reviewer/
