### 2024-07-21 #165 @editorialbot
[REVIEW]: Simulation of fracture and damage with Peridynamics.jl
**Submitting author:** @kaipartmann (<a href="http://orcid.org/0000-0002-5238-4355">Kai Partmann</a>)
**Repository:** https://github.com/kaipartmann/Peridynamics.jl
**Branch with paper.md** (empty if default branch): paper_juliacon_2023
**Version:** v0.3.3
