### 2024-07-23 ce5e8ef
Add initial setup for BANOSBMaterial
