+++ src/Peridynamics.jl
@@ -51,6 +51,7 @@ abstract type AbstractTimeSolver end
 abstract type AbstractJob end
 abstract type AbstractJobOptions end
 abstract type AbstractSystem end
+abstract type AbstractBondSystem <: AbstractSystem end
 abstract type AbstractPredefinedCrack end
 abstract type AbstractBodyChunk{S<:AbstractSystem,T<:AbstractMaterial} end
 abstract type AbstractParameterHandler <: AbstractParameterSetup end
