+++ src/discretization/bond_system.jl
@@ -4,7 +4,7 @@ struct Bond
     fail_permit::Bool
 end
 
-struct BondSystem{Correction<:AbstractCorrection} <: AbstractSystem
+struct BondSystem{Correction<:AbstractCorrection} <: AbstractBondSystem
     position::Matrix{Float64}
     volume::Vector{Float64}
     bonds::Vector{Bond}
@@ -155,7 +155,7 @@ function find_halo_points(bonds::Vector{Bond}, loc_points::UnitRange{Int})
     return halo_points
 end
 
-@inline each_bond_idx(bd::BondSystem, point_id::Int) = bd.bond_ids[point_id]
+@inline each_bond_idx(system::AbstractBondSystem, point_id::Int) = system.bond_ids[point_id]
 
 function localize!(bonds::Vector{Bond}, localizer::Dict{Int,Int})
     for i in eachindex(bonds)
@@ -165,9 +165,9 @@ function localize!(bonds::Vector{Bond}, localizer::Dict{Int,Int})
     return nothing
 end
 
-function break_bonds!(s::AbstractStorage, system::BondSystem, ch::ChunkHandler,
-                      set_a::Vector{Int}, set_b::Vector{Int})
-    s.n_active_bonds .= 0
+function break_bonds!(storage::AbstractStorage, system::AbstractBondSystem,
+                      ch::ChunkHandler, set_a::Vector{Int}, set_b::Vector{Int})
+    storage.n_active_bonds .= 0
     for point_id in each_point_idx(ch)
         for bond_id in each_bond_idx(system, point_id)
             bond = system.bonds[bond_id]
@@ -177,24 +177,24 @@ function break_bonds!(s::AbstractStorage, system::BondSystem, ch::ChunkHandler,
             neigh_in_a = in(neighbor_id, set_a)
             neigh_in_b = in(neighbor_id, set_b)
             if (point_in_a && neigh_in_b) || (point_in_b && neigh_in_a)
-                s.bond_active[bond_id] = false
+                storage.bond_active[bond_id] = false
             end
-            s.n_active_bonds[point_id] += s.bond_active[bond_id]
+            storage.n_active_bonds[point_id] += storage.bond_active[bond_id]
         end
     end
     return nothing
 end
 
-function calc_timestep_point(bd::BondSystem, params::AbstractPointParameters, point_id::Int)
+function calc_timestep_point(system::AbstractBondSystem, params::AbstractPointParameters, point_id::Int)
     dtsum = 0.0
-    for bond_id in each_bond_idx(bd, point_id)
-        bond = bd.bonds[bond_id]
-        dtsum += bd.volume[bond.neighbor] * params.bc / bond.length
+    for bond_id in each_bond_idx(system, point_id)
+        bond = system.bonds[bond_id]
+        dtsum += system.volume[bond.neighbor] * params.bc / bond.length
     end
     return sqrt(2 * params.rho / dtsum)
 end
 
-function calc_force_density!(chunk::AbstractBodyChunk{S,M}) where {S<:BondSystem,M}
+function calc_force_density!(chunk::AbstractBodyChunk{<:AbstractBondSystem})
     (; system, mat, paramsetup, storage) = chunk
     storage.b_int .= 0
     storage.n_active_bonds .= 0
@@ -204,7 +204,7 @@ function calc_force_density!(chunk::AbstractBodyChunk{S,M}) where {S<:BondSystem
     return nothing
 end
 
-@inline function calc_damage!(chunk::AbstractBodyChunk{S,M}) where {S<:BondSystem,M}
+@inline function calc_damage!(chunk::AbstractBodyChunk{<:AbstractBondSystem})
     (; n_neighbors) = chunk.system
     (; n_active_bonds, damage) = chunk.storage
[…]

[…]
