### 2024-07-23 6defaba
Add compatibility with BANOSBMaterial
