+++ Project.toml
@@ -13,7 +13,6 @@ LightXML = "9c8b4983-aa76-5018-a973-4c85ecc9e179"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 MPI = "da04e1cc-30fd-572f-bb4f-1f8673147195"
 PointNeighbors = "1c4d5385-0a27-49de-8e2c-43b175c8985c"
-Polyester = "f517fe37-dbe3-4b94-8317-1923a5111588"
 PrecompileTools = "aea7be01-6a6a-4083-8856-8a6e6704d82a"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
@@ -32,7 +31,6 @@ LightXML = "0.9"
 LinearAlgebra = "1.8"
 MPI = "0.20.0"
 PointNeighbors = "0.4"
-Polyester = "0.7.5"
 PrecompileTools = "1.0"
 Printf = "1.8"
 ProgressMeter = "1.0"
