### 2024-07-24 14e1ed6
Use threads static as parallelization backend
