### 2024-07-24 e96a43b
Update CITATION.cff
