+++ CITATION.cff
@@ -4,11 +4,13 @@ authors:
 - family-names: Partmann
   given-names: Kai
   orcid: https://orcid.org/0000-0002-5238-4355
+- family-names: Dienst
+  given-names: Manuel
+  orcid: https://orcid.org/0009-0004-9195-0112
 - family-names: Weinberg
   given-names: Kerstin
   orcid: https://orcid.org/0000-0002-2213-8401
 title: "Peridynamics.jl"
-version: 0.2.0
-doi: 10.5281/zenodo.8187478
+doi: 10.5281/zenodo.8187477
 date-released: 2023-07-26
 url: "https://github.com/kaipartmann/Peridynamics.jl"
