### 2024-07-24 3d023e8
Change compat entry for PointNeighbors
