+++ Project.toml
@@ -30,7 +30,7 @@ LibGit2 = "1.8"
 LightXML = "0.9"
 LinearAlgebra = "1.8"
 MPI = "0.20.0"
-PointNeighbors = "0.4"
+PointNeighbors = "0.4.5"
 PrecompileTools = "1.0"
 Printf = "1.8"
 ProgressMeter = "1.0"
