### 2024-07-24 ddeab3a
Allow thread nesting for reading of VTK files
