### 2024-07-25 5b6a11a
changes and additions (#124) to tutorials
