+++ docs/make.jl
@@ -34,6 +34,7 @@ LIT_TUTORIALS_IN = [
     "tutorial_wave_in_bar.jl",
     "tutorial_kalthoff-winkler_dynfrac.jl",
     "tutorial_logo.jl",
+    "tutorial_cylinder.jl",
 ]
 LIT_TUTORIALS_IN .= joinpath.(@__DIR__, "src", "literate", LIT_TUTORIALS_IN)
 Literate.markdown.(LIT_TUTORIALS_IN, LIT_MD_OUT; credit=false)
@@ -76,6 +77,7 @@ makedocs(;
             joinpath("generated", "tutorial_wave_in_bar.md"),
             joinpath("generated", "tutorial_kalthoff-winkler_dynfrac.md"),
             joinpath("generated", "tutorial_logo.md"),
+            joinpath("generated", "tutorial_cylinder.md"),
         ],
         "API Reference" => "api_reference.md",
     ],
