### 2024-07-25 66d4441
closes #93 (Consistency in explanations)
