+++ docs/src/expl_bondbased.md
@@ -5,23 +5,10 @@ The initial version of peridynamics is the bond-based (BB) formulation. [Silling
 Here, a pairwise force function $\boldsymbol{f}$ is defined and calculated for each bond of two material points, which depends on the strain of the bond and is aligned in its direction: 
 
 ```math
- \boldsymbol{f} = d^{ij} \, c \, \varepsilon^{ij} \, \boldsymbol{n} \; .
+ \boldsymbol{f} = c \, \varepsilon^{ij} \, \boldsymbol{n} \; .
 ```
 
-with
-
-| size | symbol |      unit |
-|:--------|:-------------|:------------|
-| bond failure |      $d^{ij} \in \{0;1\}$      | $[-]$ |
-| micro-modulus constant [Silling2005](@cite) |  $c = \frac{18 \cdot \kappa}{\pi \cdot \delta^4}$ | $[\frac{\mathrm{kg}}{\mathrm{m}^5\mathrm{s}^2}]$ |
-| bond strain | $\varepsilon^{ij} = \frac{l^{ij}-L^{ij}}{L^{ij}}$ |      $[-]$ |
-| bond length in $\mathcal{B}_0$ |      $L^{ij} =\left\|\boldsymbol{\Delta X}^{ij}\right\|$     | $[\mathrm{m}]$ |
-| bond length in $\mathcal{B}_t$ |      $l^{ij} =\left\|\boldsymbol{\Delta x}^{ij}\right\|$     | $[\mathrm{m}]$ |
-| direction vector |      $\boldsymbol{n} = \frac{\boldsymbol{\Delta x}^{ij}}{l^{ij}}$      | $[-]$ |
-
-In this expression, bond failure $d^{ij}$ represents whether the bond between points $i$ and $j$ is intact ($d^{ij}=1$)
-or damaged ($d^{ij}=0$). The direction vector $\boldsymbol{n}$ is oriented in the direction of the bond. 
-Furthermore, the micro-modulus constant [Silling2005](@cite)
+Here the micro-modulus constant [Silling2005](@cite)
 ```math
 c = \frac{18 \, \kappa}{\pi \, \delta^4} \;
 ```
@@ -31,9 +18,27 @@ and the strain of the bond [Silling2005a](@cite)
 ``` 
 with bond lengths $L^{ij} =\left|\boldsymbol{\Delta X}^{ij}\right|$ and $l^{ij} =\left|\boldsymbol{\Delta x}^{ij}\right|$ are used.
 
+The direction vector
+```math
+\boldsymbol{n} = \frac{\boldsymbol{\Delta x}^{ij}}{l^{ij}}
+``` 
+is oriented in the direction of the bond. 
+
 To get the resulting body forces, now the force function is integrated over the whole body:
 
 ```math
 \boldsymbol{b}^{\mathrm{int},i} = \boldsymbol{b}^{\mathrm{int}} (\boldsymbol{X} ^ {i} , t) = \int_{\mathcal{H}_i} \boldsymbol{f} \; \mathrm{d}V^j \; .
 ```
 
+| Size | Symbol |      Unit |
+|:--------|:-------------|:------------|
+| Pairwise force function | $\boldsymbol{f}$ | $\left[\frac{\mathrm{kg}}{\mathrm{m}^5\mathrm{s}^2}\right]$ |
+| Micro-modulus constant [Silling2005](@cite) |  $c$ | $\left[\frac{\mathrm{kg}}{\mathrm{m}^5\mathrm{s}^2}\right]$ |
+| Bond strain | $\varepsilon^{ij}$ |      $[-]$ |
[…]
