### 2024-07-25 3ca0b91
Update docs/src/expl_damage.md
Co-authored-by: Kai Partmann <68582683+kaipartmann@users.noreply.github.com>
