### 2024-07-25 8d8b2af
Update docs/src/expl_damage.md
Co-authored-by: Kai Partmann <68582683+kaipartmann@users.noreply.github.com>
