### 2024-07-25 73c8d98
added image of fragmenting cylinder to tutorial
