### 2024-07-25 b875964
added damage explanations (close #154)
