+++ src/physics/correspondence.jl
@@ -116,13 +116,16 @@ struct NOSBVerletStorage <: AbstractStorage
     damage::Vector{Float64}
     bond_active::Vector{Bool}
     n_active_bonds::Vector{Int}
+    rotation::Matrix{Float64}
+    left_stretch::Matrix{Float64}
 end
 
 function NOSBVerletStorage(::NOSBMaterial, ::VelocityVerlet, system::BondSystem, ch)
     n_loc_points = length(ch.loc_points)
     position = copy(system.position)
+    n_all_points = size(position, 2)
     displacement = zeros(3, n_loc_points)
-    velocity = zeros(3, n_loc_points)
+    velocity = zeros(3, n_all_points)
     velocity_half = zeros(3, n_loc_points)
     acceleration = zeros(3, n_loc_points)
     b_int = zeros(3, length(ch.point_ids))
@@ -130,14 +133,19 @@ function NOSBVerletStorage(::NOSBMaterial, ::VelocityVerlet, system::BondSystem,
     damage = zeros(n_loc_points)
     bond_active = ones(Bool, length(system.bonds))
     n_active_bonds = copy(system.n_neighbors)
+    rotation = zeros(9, n_loc_points)
+    rotation[[1, 5, 9], :] .= 1.0
+    left_stretch = zeros(9, n_loc_points)
+    left_stretch[[1, 5, 9], :] .= 1.0
     s = NOSBVerletStorage(position, displacement, velocity, velocity_half, acceleration,
-                          b_int, b_ext, damage, bond_active, n_active_bonds)
+                          b_int, b_ext, damage, bond_active, n_active_bonds, rotation,
+                          left_stretch)
     return s
 end
 
 @storage NOSBMaterial VelocityVerlet NOSBVerletStorage
 
-@loc_to_halo_fields NOSBVerletStorage :position
+@loc_to_halo_fields NOSBVerletStorage :position :velocity
 @halo_to_loc_fields NOSBVerletStorage :b_int
 
 struct NOSBRelaxationStorage <: AbstractStorage
@@ -153,13 +161,16 @@ struct NOSBRelaxationStorage <: AbstractStorage
     damage::Vector{Float64}
     bond_active::Vector{Bool}
     n_active_bonds::Vector{Int}
+    rotation::Matrix{Float64}
+    left_stretch::Matrix{Float64}
 end
 
 function NOSBRelaxationStorage(::NOSBMaterial, ::DynamicRelaxation, system::BondSystem, ch)
     n_loc_points = length(ch.loc_points)
     position = copy(system.position)
+    n_all_points = size(position, 2)
     displacement = zeros(3, n_loc_points)
-    velocity = zeros(3, n_loc_points)
+    velocity = zeros(3, n_all_points)
     velocity_half = zeros(3, n_loc_points)
     velocity_half_old = zeros(3, n_loc_points)
     b_int = zeros(3, length(ch.point_ids))
@@ -169,38 +180,53 @@ function NOSBRelaxationStorage(::NOSBMaterial, ::DynamicRelaxation, system::Bond
     damage = zeros(n_loc_points)
     bond_active = ones(Bool, length(system.bonds))
     n_active_bonds = copy(system.n_neighbors)
+    rotation = zeros(9, n_loc_points)
[…]
