+++ src/physics/bond_associated_correspondence.jl
@@ -38,6 +38,8 @@ struct BANOSBVerletStorage <: AbstractStorage
     damage::Vector{Float64}
     bond_active::Vector{Bool}
     n_active_bonds::Vector{Int}
+    rotation::Matrix{Float64}
+    left_stretch::Matrix{Float64}
 end
 
 function BANOSBVerletStorage(::BANOSBMaterial, ::VelocityVerlet,
@@ -45,7 +47,7 @@ function BANOSBVerletStorage(::BANOSBMaterial, ::VelocityVerlet,
     n_loc_points = length(ch.loc_points)
     position = copy(system.position)
     displacement = zeros(3, n_loc_points)
-    velocity = zeros(3, n_loc_points)
+    velocity = zeros(3, size(position, 2))
     velocity_half = zeros(3, n_loc_points)
     acceleration = zeros(3, n_loc_points)
     b_int = zeros(3, length(ch.point_ids))
@@ -53,14 +55,19 @@ function BANOSBVerletStorage(::BANOSBMaterial, ::VelocityVerlet,
     damage = zeros(n_loc_points)
     bond_active = ones(Bool, length(system.bonds))
     n_active_bonds = copy(system.n_neighbors)
+    rotation = zeros(9, length(system.bonds))
+    rotation[[1, 5, 9], :] .= 1.0
+    left_stretch = zeros(9, length(system.bonds))
+    left_stretch[[1, 5, 9], :] .= 1.0
     s = BANOSBVerletStorage(position, displacement, velocity, velocity_half, acceleration,
-                          b_int, b_ext, damage, bond_active, n_active_bonds)
+                            b_int, b_ext, damage, bond_active, n_active_bonds, rotation,
+                            left_stretch)
     return s
 end
 
 @storage BANOSBMaterial VelocityVerlet BANOSBVerletStorage
 
-@loc_to_halo_fields BANOSBVerletStorage :position
+@loc_to_halo_fields BANOSBVerletStorage :position :velocity
 @halo_to_loc_fields BANOSBVerletStorage :b_int
 
 struct BANOSBRelaxationStorage <: AbstractStorage
@@ -106,41 +113,48 @@ end
 
 const BANOSBStorage = Union{BANOSBVerletStorage,BANOSBRelaxationStorage}
 
-function force_density_point!(storage::BANOSBStorage, system::BondAssociatedSystem,
-                              mat::BANOSBMaterial, paramhandler::AbstractParameterHandler,
-                              i::Int)
-    params = get_params(paramhandler, i)
-    force_density_point!(storage, system, mat, params, i)
+function calc_force_density!(chunk::AbstractBodyChunk{S,M},
+                             Δt::Float64) where {S<:BondAssociatedSystem,M<:BANOSBMaterial}
+    (; system, mat, paramsetup, storage) = chunk
+    storage.b_int .= 0
+    storage.n_active_bonds .= 0
+    for point_id in each_point_idx(chunk)
+        force_density_point!(storage, system, mat, paramsetup, Δt, point_id)
+    end
     return nothing
 end
 
 function force_density_point!(storage::BANOSBStorage, system::BondAssociatedSystem,
[…]
