### 2024-08-08 2e40fa1
CompatHelper: bump compat for ThreadPinning to 1, (keep existing compat)
