+++ Project.toml
@@ -35,7 +35,7 @@ PrecompileTools = "1.0"
 Printf = "1.8"
 ProgressMeter = "1.0"
 StaticArrays = "1.0"
-ThreadPinning = "0.7"
+ThreadPinning = "0.7, 1"
 TimerOutputs = "0.5"
 WriteVTK = "1.19"
 julia = "1.9"
