+++ src/physics/correspondence.jl
@@ -1,5 +1,5 @@
 """
-    NOSBMaterial(; maxdmg, maxjacobi, corr)
+    NOSBMaterial(; maxdmg, corr)
 
 A material type used to assign the material of a [`Body`](@ref) with the local continuum
 consistent (correspondence) formulation of non-ordinary state-based peridynamics.
@@ -9,9 +9,6 @@ consistent (correspondence) formulation of non-ordinary state-based peridynamics
     exceeded, all bonds of that point are broken because the deformation gradient would then
     possibly contain `NaN` values.
     (default: `0.95`)
-- `maxjacobi::Float64`: Maximum value of the Jacobi determinant. If this value is exceeded,
-    all bonds of that point are broken.
-    (default: `1.03`)
 - `corr::Float64`: Correction factor used for zero-energy mode stabilization. The
     stabilization algorithm of Silling (2017) is used.
     (default: `100.0`)
@@ -19,13 +16,13 @@ consistent (correspondence) formulation of non-ordinary state-based peridynamics
 !!! note "Stability of fracture simulations"
     This formulation is known to be not suitable for fracture simultations without
     stabilization of the zero-energy modes. Therefore be careful when doing fracture
-    simulations and try out different paremeters for `maxdmg`, `maxjacobi`, and `corr`.
+    simulations and try out different paremeters for `maxdmg`, and `corr`.
 
 # Examples
 
 ```julia-repl
 julia> mat = NOSBMaterial()
-NOSBMaterial(maxdmg=0.95, maxjacobi=1.03, corr=100.0)
+NOSBMaterial(maxdmg=0.95, corr=100.0)
 ```
 
 ---
@@ -40,8 +37,6 @@ non-ordinary state-based peridynamics.
 # Fields
 - `maxdmg::Float64`: Maximum value of damage a point is allowed to obtain. See the
     constructor docs for more informations.
-- `maxjacobi::Float64`: Maximum value of the Jacobi determinant. See the constructor docs
-    for more informations.
 - `corr::Float64`: Correction factor used for zero-energy mode stabilization. See the
     constructor docs for more informations.
 
@@ -70,8 +65,7 @@ When specifying the `fields` keyword of [`Job`](@ref) for a [`Body`](@ref) with
 """
 Base.@kwdef struct NOSBMaterial <: AbstractBondSystemMaterial{NoCorrection}
     maxdmg::Float64 = 0.95
-    maxjacobi::Float64 = 1.03
-    corr::Float64 = 100.0
+    corr::Float64 = 0.015
 end
 
 function Base.show(io::IO, @nospecialize(mat::NOSBMaterial))
@@ -118,6 +112,9 @@ struct NOSBVerletStorage <: AbstractStorage
     n_active_bonds::Vector{Int}
     rotation::Matrix{Float64}
     left_stretch::Matrix{Float64}
+    rate_of_deformation::Matrix{Float64}
+    stress::Matrix{Float64}
+    stress_rotated::Matrix{Float64}
 end
 
 function NOSBVerletStorage(::NOSBMaterial, ::VelocityVerlet, system::BondSystem, ch)
[…]
