+++ v0.3.2/literate/tutorial_tension_precrack.jl
@@ -1,0 +1,118 @@
+# # [Tension with predefined crack](@id tutorial_tension_precrack)
+
+# Import the package:
+using Peridynamics
+# using GLMakie #src
+# using LaTeXStrings #src
+
+# First some geometrical parameters are defined.
+# These are edge length `l`, point spacing `Δx` and crack length `a`.
+# Now a cuboid body with the specified edge lengths and a thickness of one tenth thereof is
+# created using the bond-based material model.
+l, Δx, a = 1.0, 1/50, 0.5
+pos, vol = uniform_box(l, l, 0.1l, Δx)
+body = Body(BBMaterial(), pos, vol)
+
+#-
+# fig = Figure(resolution = (1000,500), figure_padding=0) #src
+# ax = Axis3(fig[1,1]; aspect = :data) #src
+# hidespines!(ax) #src
+# hidedecorations!(ax) #src
+# meshscatter!(ax, pc.position; markersize=0.5Δx, color=:grey80) #src
+# arrows!(ax, Point3[(0,-l/2,-0.08l),(-l/2,0,-0.08l),(-l/2*1.03,l/2*1.03,0)], #src
+#         Vec3[(0.1,0,0),(0,0.1,0),(0,0,0.05)], arrowsize=(0.03,0.03,0.05), #src
+#         linewidth=0.01, color=:black) #src
+# linesegments!(ax, Point3[(0,-l/2,-0.1l), (0,-l/2,-0.06l), (-l/2,0,-0.1l), #src
+#               (-l/2,0,-0.06l), (-l/2*1.03-0.02,l/2*1.03,0), #src
+#               (-l/2*1.03+0.02,l/2*1.03,0)]; linewidth=4, color=:black) #src
+# text!(ax, L"\mathbf{x}", position=Point3(0.18,-l/2,-0.08l), fontsize=30, #src
+#       align=(:center,:center), color=:black) #src
+# text!(ax, L"\mathbf{y}", position=Point3(-l/2,0.18,-0.08l), fontsize=30, #src
+#       align=(:center,:center), color=:black) #src
+# text!(ax, L"\mathbf{z}", position=Point3(-l/2*1.03,l/2*1.03,0.13), fontsize=30, #src
+#       align=(:center,:center), color=:black) #src
+# # save(joinpath(@__DIR__, "..", "assets", "tutorial_tension_1.png"), fig; px_per_unit=3) #src
+# fig #src
+
+#-
+# ![](https://github.com/kaipartmann/Peridynamics.jl/assets/68582683/ca3e724b-9dd7-417a-b3aa-bc37f81f4674) #md
+
+# The following material parameters are set:
+#
+# | material parameter | value |
+# |:--------|:-------------|
+# | Horizon $ δ $ | $3.015 \cdot Δx$ |
+# | Young's modulus $E$ | $ 210000 \, \mathrm{MPa}$ |
+# | Density $ρ$ | $ 8 \cdot 10^{-6}\,\mathrm{kg}\,\mathrm{mm}^{-3}$ |
+# | Griffith's parameter $G_c$ | $2.7 \, \mathrm{N} \, \mathrm{mm}^{-1}$ |
+δ = 3.015Δx
+material!(body; horizon=δ, E=2.1e5, rho=8e-6, Gc=2.7)
+
+# Two point sets are defined to insert a crack between them:
+point_set!(p -> p[1] ≤ -l/2+a && 0 ≤ p[2] ≤ 2δ, body, :set_a)
+point_set!(p -> p[1] ≤ -l/2+a && -2δ ≤ p[2] < 0, body, :set_b)
+precrack!(body, :set_a, :set_b)
+
+# fig = Figure(resolution = (1000,500), figure_padding=0) #src
[…]
