### 2024-08-09 c3a6d9b
Fix rotation errors; bugs with fracture simulations remain
