+++ src/physics/bond_associated_correspondence.jl
@@ -151,9 +151,9 @@ function force_density_bond!(storage::BANOSBStorage, system::BondAssociatedSyste
     end
     σ = calc_cauchy_stress(mat, params, F)
     calc_rod_and_rotation!(storage, F, Ḟ, Δt, bond_idx)
-    # R = get_rotation(storage, bond_idx)
-    # T = R * σ * R'
-    T = σ
+    R = get_rotation(storage, bond_idx)
+    T = R * σ * R'
+    # T = σ
     P = det(F) * T * inv(F)'
     PKinv = P * Kinv
 
@@ -167,8 +167,10 @@ function force_density_bond!(storage::BANOSBStorage, system::BondAssociatedSyste
 
     ωij = influence_function(mat, params, L) * storage.bond_active[bond_idx]
     tij = ωij * PKinv * ΔXij
-    temp_i = volume_fraction_factor(system, i, bond_idx) * system.volume[i]
-    temp_j = volume_fraction_factor(system, j, bond_idx) * system.volume[j]
+    # temp_i = volume_fraction_factor(system, i, bond_idx) * system.volume[i]
+    temp_i = system.volume[i]
+    # temp_j = volume_fraction_factor(system, j, bond_idx) * system.volume[j]
+    temp_j = system.volume[j]
     update_add_b_int!(storage, i, tij .* temp_j)
     update_add_b_int!(storage, j, -tij .* temp_i)
     return nothing
@@ -206,18 +208,17 @@ function calc_deformation_gradient(storage::BANOSBStorage, system::BondAssociate
     return F, Fdot, Kinv
 end
 
-function calc_first_piola_stress(F::SMatrix{3,3}, ::BANOSBMaterial,
-                                 params::BANOSBPointParameters)
-    J = det(F)
-    J < eps() && return zero(SMatrix{3,3})
-    C = F' * F
-    Cinv = inv(C)
-    S = params.G .* (I - 1 / 3 .* tr(C) .* Cinv) .* J^(-2 / 3) .+
-        params.K / 4 .* (J^2 - J^(-2)) .* Cinv
-    P = F * S
-    return P
-end
-
+# function calc_first_piola_stress(F::SMatrix{3,3}, ::BANOSBMaterial,
+#                                  params::BANOSBPointParameters)
+#     J = det(F)
+#     J < eps() && return zero(SMatrix{3,3})
+#     C = F' * F
+#     Cinv = inv(C)
+#     S = params.G .* (I - 1 / 3 .* tr(C) .* Cinv) .* J^(-2 / 3) .+
+#         params.K / 4 .* (J^2 - J^(-2)) .* Cinv
+#     P = F * S
+#     return P
+# end
 
 function calc_cauchy_stress(mat::BANOSBMaterial, params::BANOSBPointParameters, F::SMatrix{3,3})
     # ---- Neo-Hookean model
@@ -236,7 +237,6 @@ function calc_cauchy_stress(mat::BANOSBMaterial, params::BANOSBPointParameters, 
     #      volumetric response:
     J = det(F)
[…]
