### 2024-08-09 375868d
Renamed NOSBMaterial -> CCMaterial
