+++ src/physics/correspondence.jl
@@ -1,5 +1,5 @@
 """
-    NOSBMaterial(; maxdmg, maxjacobi, corr)
+    CCMaterial(; maxdmg, maxjacobi, corr)
 
 A material type used to assign the material of a [`Body`](@ref) with the local continuum
 consistent (correspondence) formulation of non-ordinary state-based peridynamics.
@@ -24,14 +24,14 @@ consistent (correspondence) formulation of non-ordinary state-based peridynamics
 # Examples
 
 ```julia-repl
-julia> mat = NOSBMaterial()
-NOSBMaterial(maxdmg=0.95, maxjacobi=1.03, corr=100.0)
+julia> mat = CCMaterial()
+CCMaterial(maxdmg=0.95, maxjacobi=1.03, corr=100.0)
 ```
 
 ---
 
 ```julia
-NOSBMaterial
+CCMaterial
 ```
 
 Material type for the local continuum consistent (correspondence) formulation of
@@ -46,7 +46,7 @@ non-ordinary state-based peridynamics.
     constructor docs for more informations.
 
 # Allowed material parameters
-When using [`material!`](@ref) on a [`Body`](@ref) with `NOSBMaterial`, then the following
+When using [`material!`](@ref) on a [`Body`](@ref) with `CCMaterial`, then the following
 parameters are allowed:
 - `horizon::Float64`: Radius of point interactions
 - `rho::Float64`: Density
@@ -57,7 +57,7 @@ parameters are allowed:
 
 # Allowed export fields
 When specifying the `fields` keyword of [`Job`](@ref) for a [`Body`](@ref) with
-`NOSBMaterial`, the following fields are allowed:
+`CCMaterial`, the following fields are allowed:
 - `position::Matrix{Float64}`: Position of each point
 - `displacement::Matrix{Float64}`: Displacement of each point
 - `velocity::Matrix{Float64}`: Velocity of each point
@@ -68,13 +68,13 @@ When specifying the `fields` keyword of [`Job`](@ref) for a [`Body`](@ref) with
 - `damage::Vector{Float64}`: Damage of each point
 - `n_active_bonds::Vector{Int}`: Number of intact bonds of each point
 """
-Base.@kwdef struct NOSBMaterial <: AbstractBondSystemMaterial{NoCorrection}
+Base.@kwdef struct CCMaterial <: AbstractBondSystemMaterial{NoCorrection}
     maxdmg::Float64 = 0.95
     maxjacobi::Float64 = 1.03
     corr::Float64 = 100.0
 end
 
-function Base.show(io::IO, @nospecialize(mat::NOSBMaterial))
+function Base.show(io::IO, @nospecialize(mat::CCMaterial))
     print(io, typeof(mat))
     print(io, msg_fields_in_brackets(mat))
     return nothing
@@ -94,7 +94,7 @@ struct NOSBPointParameters <: AbstractPointParameters
     bc::Float64
 end
 
-function NOSBPointParameters(::NOSBMaterial, p::Dict{Symbol,Any})
+function NOSBPointParameters(::CCMaterial, p::Dict{Symbol,Any})
     δ = get_horizon(p)
     rho = get_density(p)
     E, nu, G, K, λ, μ = get_elastic_params(p)
@@ -103,7 +103,7 @@ function NOSBPointParameters(::NOSBMaterial, p::Dict{Symbol,Any})
     return NOSBPointParameters(δ, rho, E, nu, G, K, λ, μ, Gc, εc, bc)
 end
 
-@params NOSBMaterial NOSBPointParameters
+@params CCMaterial NOSBPointParameters
 
 struct NOSBVerletStorage <: AbstractStorage
     position::Matrix{Float64}
@@ -118,7 +118,7 @@ struct NOSBVerletStorage <: AbstractStorage
[…]
