+++ test/integration/symmetry_correspondence.jl
@@ -1,4 +1,4 @@
-@testitem "symmetry NOSBMaterial" begin
+@testitem "symmetry CCMaterial" begin
     # simulation
     Δx = 0.2
     width = 1
@@ -8,7 +8,7 @@
     pos = hcat(([x;y;z] for x in grid for y in grid for z in grid)...)
     n_points = size(pos, 2)
     vol = fill(Δx^3, n_points)
-    body = Body(NOSBMaterial(), pos, vol)
+    body = Body(CCMaterial(), pos, vol)
     failure_permit!(body, false)
     material!(body, horizon=3.015Δx, rho=7850, E=210e9, nu=0.25, Gc=1)
     point_set!(z -> z > width/2 - 0.6Δx, body, :set_a)
@@ -57,7 +57,7 @@
     rm(temp_root; recursive=true, force=true)
 end
 
-@testitem "symmetry NOSBMaterial DynamicRelaxation" begin
+@testitem "symmetry CCMaterial DynamicRelaxation" begin
     # simulation
     Δx = 0.2
     width = 1
@@ -67,7 +67,7 @@ end
     pos = hcat(([x;y;z] for x in grid for y in grid for z in grid)...)
     n_points = size(pos, 2)
     vol = fill(Δx^3, n_points)
-    body = Body(NOSBMaterial(), pos, vol)
+    body = Body(CCMaterial(), pos, vol)
     failure_permit!(body, false)
     material!(body, horizon=3.015Δx, rho=7850, E=210e9, nu=0.25, Gc=1)
     point_set!(z -> z > width/2 - 0.6Δx, body, :set_a)
