### 2024-08-09 d31fd31
Renaming NOSBPointParameters -> CCPointParameters
