### 2024-08-09 11a9259
Renaming NOSBStorage -> CCStorage
