### 2024-08-09 4b3e89a
Remove `maxjacobi` field
