### 2024-08-09 c77cd11
Added time and step as args to calculation of force density
