+++ src/physics/correspondence.jl
@@ -22,7 +22,7 @@ consistent (correspondence) formulation of non-ordinary state-based peridynamics
 
 ```julia-repl
 julia> mat = CCMaterial()
-CCMaterial(maxdmg=0.85, maxjacobi=1.03, corr=100.0)
+CCMaterial(maxdmg=0.85, corr=100.0)
 ```
 
 ---
@@ -63,18 +63,25 @@ When specifying the `fields` keyword of [`Job`](@ref) for a [`Body`](@ref) with
 - `damage::Vector{Float64}`: Damage of each point
 - `n_active_bonds::Vector{Int}`: Number of intact bonds of each point
 """
-struct CCMaterial <: AbstractBondSystemMaterial{NoCorrection}
+struct CCMaterial{CM,SI} <: AbstractCorrespondenceMaterial{CM,SI,NoCorrection}
+    constitutive_model::CM
+    stress_integration::SI
     maxdmg::Float64
     corr::Float64
+    function CCMaterial(cm::CM, si::SI, maxdmg::Real, corr::Real) where {CM,SI}
+        return new{CM,SI}(cm, si, maxdmg, corr)
+    end
 end
 
-function CCMaterial(; maxdmg=0.85, corr=100.0)
-    return CCMaterial(maxdmg, corr)
+function CCMaterial(constitutive_model::AbstractConstitutiveModel=NeoHookeNonlinear();
+                    stress_integration::AbstractStressIntegration=FlanaganTaylorRotation(),
+                    maxdmg::Real=0.85, corr::Real=100.0)
+    return CCMaterial(constitutive_model, stress_integration, maxdmg, corr)
 end
 
 function Base.show(io::IO, @nospecialize(mat::CCMaterial))
     print(io, typeof(mat))
-    print(io, msg_fields_in_brackets(mat))
+    print(io, msg_fields_in_brackets(mat, (:maxdmg, :corr)))
     return nothing
 end
 
@@ -114,28 +121,36 @@ struct CCVerletStorage <: AbstractStorage
     damage::Vector{Float64}
     bond_active::Vector{Bool}
     n_active_bonds::Vector{Int}
+    rotation::Matrix{Float64}
+    left_stretch::Matrix{Float64}
+    stress::Matrix{Float64}
 end
 
-function CCVerletStorage(::CCMaterial, ::VelocityVerlet, system::BondSystem, ch)
+function CCVerletStorage(mat::CCMaterial, ::VelocityVerlet, system::BondSystem, ch)
     n_loc_points = length(ch.loc_points)
+    n_all_points = length(ch.point_ids)
     position = copy(system.position)
     displacement = zeros(3, n_loc_points)
-    velocity = zeros(3, n_loc_points)
+    velocity = zeros(3, n_all_points)
     velocity_half = zeros(3, n_loc_points)
     acceleration = zeros(3, n_loc_points)
-    b_int = zeros(3, length(ch.point_ids))
+    b_int = zeros(3, n_all_points)
     b_ext = zeros(3, n_loc_points)
     damage = zeros(n_loc_points)
     bond_active = ones(Bool, length(system.bonds))
     n_active_bonds = copy(system.n_neighbors)
+    rotation = init_rotation(mat, n_loc_points)
+    left_stretch = init_left_stretch(mat, n_loc_points)
+    stress = zeros(9, n_loc_points)
     s = CCVerletStorage(position, displacement, velocity, velocity_half, acceleration,
-                        b_int, b_ext, damage, bond_active, n_active_bonds)
+                        b_int, b_ext, damage, bond_active, n_active_bonds, rotation,
+                        left_stretch, stress)
     return s
 end
 
[…]
