### 2024-08-14 e855a57
Export SaintVenantKirchhoff model
