+++ Project.toml
@@ -17,6 +17,8 @@ PrecompileTools = "aea7be01-6a6a-4083-8856-8a6e6704d82a"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
+TensorOperations = "6aa20fa7-93e2-5fca-9bc0-fbd0db3c71a2"
+Tensors = "48a634ad-e948-5137-8d70-aa71f2a747f4"
 ThreadPinning = "811555cd-349b-4f26-b7bc-1f208b848042"
 TimerOutputs = "a759f4b9-e2f1-59dc-863e-4aeb61b1ea8f"
 WriteVTK = "64499a7a-5c06-52f2-abe2-ccb03c286192"
@@ -35,6 +37,8 @@ PrecompileTools = "1.0"
 Printf = "1.8"
 ProgressMeter = "1.0"
 StaticArrays = "1.0"
+TensorOperations = "5"
+Tensors = "1.16"
 ThreadPinning = "0.7, 1"
 TimerOutputs = "0.5"
 WriteVTK = "1.19"
