### 2024-08-15 ff63b43
Reduced allocations
