### 2024-08-17 kaipartmann/Peridynamics.jl#167 @ranocha
JuliaCon review: errors
This is part of the review https://github.com/JuliaCon/proceedings-review/issues/165

I installed the package on Julia 1.10 and started the tests. They errored with several messages like

```
set_progress_bars!: Log Test Failed at ~/.julia/packages/Peridynamics/SePS6/test/auxiliary/test_logs.jl:15
  Expression: Peridynamics.set_progress_bars!()
  Log Pattern: (:warn,)
  Captured Logs:
```

and

```
┌ Error: Multibody simulations with MPI are not yet implemented!
└ @ Peridynamics ~/.julia/packages/Peridynamics/SePS6/src/discretization/multibody_setup.jl:122
application called MPI_Abort(MPI_COMM_WORLD, 1) - process 0
[unset]: PMIU_write error; fd=-1 buf=:cmd=abort exitcode=1 message=application called MPI_Abort(MPI_COMM_WORLD, 1) - process 0
:
system msg for write_line failure : Bad file descriptor
ERROR: Package Peridynamics errored during testing
```

Please fix the tests so that they can be run locally.
