### 2024-08-17 kaipartmann/Peridynamics.jl#168 @ranocha
JuliaCon review
This is part of the review https://github.com/JuliaCon/proceedings-review/issues/165

Thanks a lot for your contributions 🙂 Please find below some comments based on the review checklist:

- [x] Review item "Does the full list of paper authors seem appropriate and complete?":
      Based on the [contributors overview](https://github.com/kaipartmann/Peridynamics.jl/graphs/contributors), it is questionable whether Kerstin Weinberg can be an author of this publication - one commit with two lines changes in the README.md. There is also another contributor [oldninja](https://github.com/oldninja) with more commits.
      Please clarify this aspect (e.g., by extended guidance and mentorship by her, pair programming, ...) and note that basically the [JOSS policy](https://joss.readthedocs.io/en/latest/submitting.html#authorship) applies.
- [x] Can you please make the legend of Figure 2 bigger to make it easier to read?
- [x] Could you please clarify the audience of you package in the statement of need section a bit? Is it mostly interesting for people developing new methods as a playground or for practitioners interested in high-performance computations with default methods?
- [x] There are performance claims in the paper (section 3) without any details or further references. This conflicts with one of the items in the review checklist.
- [x] I get a 404 error when clicking on the "Report a Vulnerability" in SECURITY.md.
- [x] There are no figures in https://kaipartmann.github.io/Peridynamics.jl/stable/generated/tutorial_wave_in_bar/ - why?
- [ ] Please fix https://github.com/kaipartmann/Peridynamics.jl/issues/167
