### 2024-08-17 #165 @ranocha
I have created the issues https://github.com/kaipartmann/Peridynamics.jl/issues/167 and https://github.com/kaipartmann/Peridynamics.jl/issues/168 for this review. I think the paper can be accepted after major revisions.
