### 2024-08-17 #165 @luraess
Thanks @ranocha for providinga review of the article draft. A minor remark wrt the "statement of need"; we are currently revising the guidelines for that and do no longer require an explicit section about it. We leave it up to the authors to state and motivate where appropriate the need for their work.
