### 2024-08-17 e2b7421
Claim tests that for now do not work with MPI
