### 2024-08-21 kaipartmann/Peridynamics.jl#167 @kaipartmann
Hello @ranocha,
Thank you very much for your review and for finding these errors! I could not reproduce all the errors, but I think the changes made in https://github.com/kaipartmann/Peridynamics.jl/pull/169 should fix them. Please test the code in the branch `juliacon_review_errors` and check if my changes make the tests work on your machine. Please also see #170 for the bug that should be the root cause of the errors.
