### 2024-08-22 de7b657
add functions and tests for round spheres
