+++ src/discretization/point_generators.jl
@@ -120,3 +120,231 @@ function uniform_sphere(diameter::Real, ΔX0::Real; center_x::Real=0, center_y::
     isapprox(center_z, 0; atol=eps()) || (position[3, :] .+= center_z)
     return position, volume
 end
+
+
+
+# ###################################################################
+
+"""
+TODO
+"""
+function sphere_shape(diameter::Real, ΔX0::Real; center_position=(0, 0, 0), tlsph=false)
+    radius = diameter / 2
+    NDIMS = length(center_position)
+    coordinates = sphere_shape_coords(ΔX0, radius,
+                                      SVector{NDIMS}(center_position),
+                                      tlsph)
+    n_points = size(coordinates, 2)
+    volumes = zeros(n_points)
+    volumes .= 4/3 * π * radius^3 / n_points
+
+    return coordinates, volumes
+end
+
+"""
+TODO
+"""
+# Richtung der längsachse ändern?
+function cylinder(diameter::Real, height::Real, ΔX0::Real; center_position=(0, 0, 0), tlsph=false)
+    radius = diameter / 2
+
+    xy = sphere_shape_coords(ΔX0, radius,
+                             SVector{2}((center_position[1], center_position[2])), tlsph)
+    _z = range(-height/2, height/2, step=ΔX0)
+    z = _z .- sum(_z) / length(_z) .+ center_position[3]
+    n_layers = length(z)
+    n_points_per_layer = size(xy, 2)
+    n_points = n_points_per_layer * n_layers
+    coordinates = zeros(3, n_points)
+    a, b = 1, 0
+    for layer_id in 1:n_layers
+        b += n_points_per_layer
+        coordinates[1:2, a:b] .= xy
+        coordinates[3, a:b] .= z[layer_id]
+        a += n_points_per_layer
+    end
+    volumes = zeros(n_points)
+    volumes .= 2 * π * radius^2 * height / n_points
+    return coordinates, volumes
+end
+
+##########################################
+
+function sphere_shape_coords(particle_spacing, radius, center, tlsph)
+    if tlsph
+        # Just create a sphere that is 0.5 particle spacing larger
+        radius += 0.5particle_spacing
+    end
+
+    # Each layer has thickness `particle_spacing`
+    n_layers = round(Int, radius / particle_spacing)
+
+    if n_layers < 1
+        # Just return one particle at the center
+        return collect(reshape(center, (length(center), 1)))
+    end
+
+    # Same as above, which puts the inner radius between 0 and `particle_spacing`
+    inner_radius = max(0.0, radius - n_layers * particle_spacing + 0.5particle_spacing)
+
+    coords = zeros(length(center), 0)
+
+    for layer in 0:(n_layers - 1)
+        sphere_coords = round_sphere(particle_spacing,
+                                     inner_radius + layer * particle_spacing, center)
+        coords = hcat(coords, sphere_coords)
+    end
+
+    return coords
+end
+
+function round_sphere(particle_spacing, radius, center::SVector{2})
+
+    n_particles = round(Int, 2pi * radius / particle_spacing)
+
+    if n_particles <= 2
+        # 2 or less particles produce weird, asymmetric results.
+        # Just return one particle at the center.
+        return collect(reshape(center, (2, 1)))
+    end
+
+    # Remove the last particle at 2pi, which overlaps with the first at 0
+    t = LinRange(0, 2pi, n_particles + 1)[1:(end - 1)]
+
+    particle_coords = Array{Float64, 2}(undef, 2, length(t))
+
+    for i in axes(particle_coords, 2)
+        particle_coords[:, i] = center + radius * SVector(cos(t[i]), sin(t[i]))
+    end
+
+    return particle_coords
+end
+
+function round_sphere(particle_spacing, radius, center::SVector{3})
[…]
