### 2024-08-22 1fbe0d4
changed embedding of mp4 in tutorials to make them accessible in all browsers
