### 2024-08-29 598f8e1
employ consistent names for point generator functions
