+++ test/discretization/test_point_generators.jl
@@ -34,8 +34,8 @@ end
     @test vol == fill(0.125, 8)
 end
 
-@testitem "sphere_shape" begin
-    pos, vol = Peridynamics.sphere_shape(1.6, 0.5)
+@testitem "round_sphere" begin
+    pos, vol = Peridynamics.round_sphere(1.6, 0.5)
     @test pos ≈ [
         0.0 0.0 0.47533372629569887 0.23766686314784938 -0.23766686314784954 -0.47533372629569887 -0.23766686314784924 0.2376668631478495 0.492143194265497 0.30684626267871895 -0.10951216322820224 -0.4434056965832651 -0.4434056965832652 -0.10951216322820236 0.30684626267871884 0.0;
         0.0 0.0 0.0 0.4116510822475945 0.41165108224759445 5.821159264348117e-17 -0.41165108224759456 -0.41165108224759445 0.0 0.3847730431591451 0.47980413687975465 0.2135329293091727 -0.2135329293091726 -0.47980413687975465 -0.3847730431591452 0.0;
@@ -43,7 +43,7 @@ end
     ]
     @test vol == fill(0.13404128655316452, 16)
 
-    pos, vol = Peridynamics.sphere_shape(1, 0.5)
+    pos, vol = Peridynamics.round_sphere(1, 0.5)
     @test pos ≈ [
         0.25 -0.125 -0.125;
         0.0 0.21650635094610968 -0.21650635094610968;
@@ -51,7 +51,7 @@ end
     ]
     @test vol == fill(0.17453292519943295, 3)
 
-    pos, vol = Peridynamics.sphere_shape(0.8, 0.5)
+    pos, vol = Peridynamics.round_sphere(0.8, 0.5)
     @test pos ≈ [
         0.0
         0.0
@@ -59,7 +59,7 @@ end
     ]
     @test vol == fill(0.26808257310632905, 1)
 
-    pos, vol = Peridynamics.sphere_shape(1.6, 0.5; center_position=(1, 2, 3))
+    pos, vol = Peridynamics.round_sphere(1.6, 0.5; center_position=(1, 2, 3))
     @test pos ≈ [
         1.0 1.0 1.475333726295699 1.2376668631478493 0.7623331368521504 0.5246662737043011 0.7623331368521508 1.2376668631478496 1.492143194265497 1.306846262678719 0.8904878367717978 0.5565943034167349 0.5565943034167349 0.8904878367717977 1.306846262678719 1.0;
         2.0 2.0 2.0 2.4116510822475945 2.4116510822475945 2.0 1.5883489177524055 1.5883489177524055 2.0 2.384773043159145 2.4798041368797548 2.2135329293091726 1.7864670706908274 1.5201958631202452 1.6152269568408548 2.0;
@@ -67,7 +67,7 @@ end
     ]
     @test vol == fill(0.13404128655316452, 16)
 
-    pos, vol = Peridynamics.sphere_shape(1, 0.5; center_position=(1, 2, 3))
+    pos, vol = Peridynamics.round_sphere(1, 0.5; center_position=(1, 2, 3))
     @test pos ≈ [
         1.25 0.875 0.875;
         2.0 2.21650635094611 1.7834936490538904;
@@ -75,7 +75,7 @@ end
     ]
     @test vol == fill(0.17453292519943295, 3)
 
-    pos, vol = Peridynamics.sphere_shape(0.8, 0.5; center_position=(1, 2, 3))
+    pos, vol = Peridynamics.round_sphere(0.8, 0.5; center_position=(1, 2, 3))
     @test pos ≈ [
         1.0
         2.0
@@ -84,8 +84,8 @@ end
[…]
