### 2024-08-29 867d33c
added function uniform_cylinder and related tests to point generators
