### 2024-08-29 4ef7cd8
added docstrings for new point generator functions
