+++ test/discretization/test_point_generators.jl
@@ -45,9 +45,9 @@ end
 
     pos, vol = Peridynamics.uniform_cylinder(1, 2, 0.5; center_position=(1, 2, 1))
     @test pos ≈ [
-        0.75   1.25  0.75   1.25  0.75   1.25  0.75   1.25  0.75   1.25  0.75  1.25  0.75   1.25  0.75  1.25
-        1.75  1.75   2.25   2.25  1.75  1.75   2.25   2.25  1.75  1.75   2.25  2.25  1.75  1.75   2.25  2.25
-        0.25  0.25  0.25  0.25  0.75  0.75  0.75  0.75   1.25   1.25   1.25  1.25   1.75   1.75   1.75  1.75
+        0.75  1.25  0.75  1.25  0.75  1.25  0.75  1.25  0.75  1.25  0.75  1.25  0.75  1.25  0.75  1.25
+        1.75  1.75  2.25  2.25  1.75  1.75  2.25  2.25  1.75  1.75  2.25  2.25  1.75  1.75  2.25  2.25
+        0.25  0.25  0.25  0.25  0.75  0.75  0.75  0.75  1.25  1.25  1.25  1.25  1.75  1.75  1.75  1.75
     ]
     @test vol == fill(0.125, 16)
 end
