### 2024-09-02 kaipartmann/Peridynamics.jl#167 @kaipartmann
Hello @ranocha, thank you very much for your help! So the remaining errors are an issue with these tests that run MPI like suggested in the [MPI.jl docs](https://juliaparallel.org/MPI.jl/latest/usage/#Writing-MPI-tests):
https://github.com/kaipartmann/Peridynamics.jl/blob/main/test/integration/mpi_threads_comparison.jl

Can you run this code in the same environment?
```julia-repl
julia> mpi_cmd = """
       using Peridynamics
       using Peridynamics: mpi_rank, mpi_run
       @mpiroot sleep(0.5)
       @show mpi_rank(), mpi_run()
       """
"using Peridynamics\nusing Peridynamics: mpi_rank, mpi_run\n@mpiroot sleep(0.5)\n@show mpi_rank(), mpi_run()\n"

julia> run(`$(Peridynamics.MPI.mpiexec()) -n 2 $(Base.julia_cmd()) --project -e $(mpi_cmd)`);
(mpi_rank(), mpi_run()) = (1, true)
(mpi_rank(), mpi_run()) = (0, true)
```

If not, maybe it could be an issue with the MPI installation.
