### 2024-09-03 bc1cc10
Increase legend of Fig. 2
