+++ paper/paper.tex
@@ -62,7 +62,7 @@ A recent development in this regard is continuum-kinematics-inspired peridynamic
 Another peridynamic formulation is the local continuum consistent correspondence formulation of non-ordinary state-based peridynamics, where an elastic model from the classical local material theory can be used to calculate the internal force density.
 
 \begin{figure}
-\centerline{\includegraphics[width=0.8\linewidth]{tensile_test.png}}
+\centerline{\includegraphics[width=0.9\linewidth]{tensile_test.png}}
 \caption{Fracture simulation of tensile tension test with a crack propagating in the middle of the specimen}
 \label{fig:tensiletest}
 \end{figure}
