+++ paper/ref.bib
@@ -112,12 +112,14 @@
 }
 
 @article{Partmann2024AAM,
-  author  = {Kai Partmann and Manuel Dienst and Kerstin Weinberg},
-  journal = {submitted to Archive of Applied Mechanics},
-  number  = {},
-  title   = {Peridynamic computations of wave propagation and reflection at material interfaces},
-  volume  = {},
-  year    = {2024}
+  author    = {Partmann, Kai and Dienst, Manuel and Weinberg, Kerstin},
+  journal   = {Archive of Applied Mechanics},
+  title     = {Peridynamic computations of wave propagation and reflection at material interfaces},
+  year      = {2024},
+  issn      = {1432-0681},
+  month     = jul,
+  doi       = {10.1007/s00419-024-02646-x},
+  publisher = {Springer Science and Business Media LLC},
 }
 
 @article{Partmann2024PAMM,
