### 2024-09-03 30674ef
Updated the perspectives section
