### 2024-09-03 c88e01d
Update the statement of need section to clarify the audience of the package
