### 2024-09-03 01f5bd7
Minor shortifications
