### 2024-09-03 kaipartmann/Peridynamics.jl#168 @kaipartmann
Hello @ranocha,

Thank you very much for your review! I pushed some changes to the paper branch and addressed most of the points you mentioned.

I have some questions regarding the open points.

Regarding the first bullet point, Kerstin Weinberg provided extended guidance and support throughout the development of the Julia package, offering regular feedback on its direction and plans. Extended pair programming sessions were also used, enhancing our debugging approach and overall code quality.
 Is adding an author's contribution section to the paper necessary to clarify this aspect, or is this clarification just needed for the review process?

I checked the link in SECURITY.md with different machines, and it worked for me. I compared it with the link in
https://github.com/trixi-framework/Trixi.jl/blob/5c9332b944fb37d2432ea9e14c7d8fc77ce8ee24/SECURITY.md?plain=1#L18,
and it has the same URL structure.
On which browser/device did you check the link?
