### 2024-09-03 #165 @kaipartmann
Hello together,
I made some changes addressing the review comments in https://github.com/kaipartmann/Peridynamics.jl/issues/168.
Again, thank you for the valuable feedback!
@ranocha, @HaoZeke, please let me know if you have further remarks on these changes.
