### 2024-09-05 90f8ff2
Correct formatting of function uniform_cylinder
Co-authored-by: Kai Partmann <68582683+kaipartmann@users.noreply.github.com>
