### 2024-09-05 a83455a
Added reference to `TrixiParticles.jl` to docstring of round_sphere
Co-authored-by: Kai Partmann <68582683+kaipartmann@users.noreply.github.com>
