### 2024-09-05 b2b62f7
exported new point generator functions and adapted tests
