+++ test/discretization/test_point_generators.jl
@@ -35,7 +35,7 @@ end
 end
 
 @testitem "uniform_cylinder" begin
-    pos, vol = Peridynamics.uniform_cylinder(1, 2, 0.5)
+    pos, vol = uniform_cylinder(1, 2, 0.5)
     @test pos ≈ [
         -0.25   0.25  -0.25   0.25  -0.25   0.25  -0.25   0.25  -0.25   0.25  -0.25  0.25  -0.25   0.25  -0.25  0.25
         -0.25  -0.25   0.25   0.25  -0.25  -0.25   0.25   0.25  -0.25  -0.25   0.25  0.25  -0.25  -0.25   0.25  0.25
@@ -43,7 +43,7 @@ end
     ]
     @test vol == fill(0.125, 16)
 
-    pos, vol = Peridynamics.uniform_cylinder(1, 2, 0.5; center_position=(1, 2, 1))
+    pos, vol = uniform_cylinder(1, 2, 0.5; center_position=(1, 2, 1))
     @test pos ≈ [
         0.75  1.25  0.75  1.25  0.75  1.25  0.75  1.25  0.75  1.25  0.75  1.25  0.75  1.25  0.75  1.25
         1.75  1.75  2.25  2.25  1.75  1.75  2.25  2.25  1.75  1.75  2.25  2.25  1.75  1.75  2.25  2.25
@@ -54,7 +54,7 @@ end
 
 
 @testitem "round_sphere" begin
-    pos, vol = Peridynamics.round_sphere(1.6, 0.5)
+    pos, vol = round_sphere(1.6, 0.5)
     @test pos ≈ [
         0.0 0.0 0.47533372629569887 0.23766686314784938 -0.23766686314784954 -0.47533372629569887 -0.23766686314784924 0.2376668631478495 0.492143194265497 0.30684626267871895 -0.10951216322820224 -0.4434056965832651 -0.4434056965832652 -0.10951216322820236 0.30684626267871884 0.0;
         0.0 0.0 0.0 0.4116510822475945 0.41165108224759445 5.821159264348117e-17 -0.41165108224759456 -0.41165108224759445 0.0 0.3847730431591451 0.47980413687975465 0.2135329293091727 -0.2135329293091726 -0.47980413687975465 -0.3847730431591452 0.0;
@@ -62,7 +62,7 @@ end
     ]
     @test vol == fill(0.13404128655316452, 16)
 
-    pos, vol = Peridynamics.round_sphere(1, 0.5)
+    pos, vol = round_sphere(1, 0.5)
     @test pos ≈ [
         0.25 -0.125 -0.125;
         0.0 0.21650635094610968 -0.21650635094610968;
@@ -70,7 +70,7 @@ end
     ]
     @test vol == fill(0.17453292519943295, 3)
 
-    pos, vol = Peridynamics.round_sphere(0.8, 0.5)
+    pos, vol = round_sphere(0.8, 0.5)
     @test pos ≈ [
         0.0
         0.0
@@ -78,7 +78,7 @@ end
     ]
     @test vol == fill(0.26808257310632905, 1)
 
-    pos, vol = Peridynamics.round_sphere(1.6, 0.5; center_position=(1, 2, 3))
+    pos, vol = round_sphere(1.6, 0.5; center_position=(1, 2, 3))
     @test pos ≈ [
         1.0 1.0 1.475333726295699 1.2376668631478493 0.7623331368521504 0.5246662737043011 0.7623331368521508 1.2376668631478496 1.492143194265497 1.306846262678719 0.8904878367717978 0.5565943034167349 0.5565943034167349 0.8904878367717977 1.306846262678719 1.0;
[…]
