### 2024-09-05 kaipartmann/Peridynamics.jl#168 @ranocha
> Regarding the first bullet point, Kerstin Weinberg provided extended guidance and support throughout the development of the Julia package, offering regular feedback on its direction and plans. Extended pair programming sessions were also used, enhancing our debugging approach and overall code quality.
> Is adding an author's contribution section to the paper necessary to clarify this aspect, or is this clarification just needed for the review process?

It's required for the review process. Your answer is fine with me - that's why I have suggested several options to justify this. Its's just part of the (JOSS) review process, which is a bit different from standard journals (due to the focus on code).
