### 2024-09-05 kaipartmann/Peridynamics.jl#168 @ranocha
Please remember to release a new version of the software including your recent improvements - and to merge/rebase your paper branch accordingly so that it contains the full software (see https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements)
