### 2024-09-10 kaipartmann/Peridynamics.jl#168 @kaipartmann
> > I checked the link in SECURITY.md with different machines, and it worked for me. I compared it with the link in
> > https://github.com/trixi-framework/Trixi.jl/blob/5c9332b944fb37d2432ea9e14c7d8fc77ce8ee24/SECURITY.md?plain=1#L18,
> > and it has the same URL structure.
>
> Did you try that while not being logged-in as an admin of this repo?

Yes, within a private browser window and on different machines. Then it is necessary to sign in to submit a security vulnerability.
