### 2024-09-10 kaipartmann/Peridynamics.jl#168 @kaipartmann
> Please remember to release a new version of the software including your recent improvements - and to merge/rebase your paper branch accordingly so that it contains the full software (see https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements)

Thank you for this reminder! I will submit a patch release with all changes after #167 is fixed.
