### 2024-09-10 kaipartmann/Peridynamics.jl#167 @kaipartmann
@ranocha Can you run this code? This is just the hello world example of the MPI.jl package:
```julia-repl
julia> mpi_cmd = raw"""
       using Peridynamics
       using Peridynamics: MPI
       MPI.Init()
       comm = MPI.COMM_WORLD
       print("Hello world, I am rank $(MPI.Comm_rank(comm)) of $(MPI.Comm_size(comm))\n")
       MPI.Barrier(comm)
       """;

julia> run(`$(Peridynamics.MPI.mpiexec()) -n 2 $(Base.julia_cmd()) --project -e $(mpi_cmd)`);
Hello world, I am rank 0 of 2
Hello world, I am rank 1 of 2
```
If not, then please add MPI
```
(jl_vbKyJL) pkg> add MPI
```
and try again.
Thank you very much for your ongoing help and support!
