### 2024-09-11 kaipartmann/Peridynamics.jl#168 @ranocha
I still get
