### 2024-09-11 kaipartmann/Peridynamics.jl#168 @kaipartmann
Hm, that is really interesting...🧐
I have just tested it again with 3 of my colleagues on Linux, Windows and Mac and it worked for them.
If it has anything to do with it, I also have enabled “Private vulnerability reporting” and some other advanced settings, maybe that solved it for you.
Please check again, and also check if the green button "Report a vulnerability" is working. And please let me know which URL this button has on your machine, because this should be the same as in `SECURITY.md`.
