### 2024-09-11 kaipartmann/Peridynamics.jl#168 @ranocha
I'll close this in favor of #167
