### 2024-09-11 kaipartmann/Peridynamics.jl#167 @ranocha
I also get some additional errors earlier, e.g.,

```
log_create_data_handler: Test Failed at ~/.julia/packages/Peridynamics/SePS6/test/auxiliary/test_logs.jl:46
  Expression: Peridynamics.progress_bars() == false
   Evaluated: true == false
```
