### 2024-09-11 kaipartmann/Peridynamics.jl#167 @kaipartmann
> running tests with Peridynamics v0.3.2 locally

So did you try to run the tests using the version in the `juliacon_review_errors` branch?

However, this seems to be related to #170.
Do you have an environment parameter `MPI_LOCALRANKID` defined at your system? If yes, then [`mpi_run_initial_check`](https://github.com/kaipartmann/Peridynamics.jl/blob/e2b7421bf911f05a9b3cec9a03990224653aa3ad/src/auxiliary/mpi.jl#L63-L77) would always decide to use the MPI backend, which could result in the errors.

Thank you again for your patience and ongoing support regarding these issues! It really helps to improve the package!
