### 2024-09-11 kaipartmann/Peridynamics.jl#167 @ranocha
If you think your PR fixes the issue, could you please make it available in the main release? It's kind of annoying for me to test a PR branch.

I don't have `MPI_LOCALRANKID` defined
