### 2024-09-11 kaipartmann/Peridynamics.jl#167 @kaipartmann
@ranocha The PR has now been merged with the main branch. I'm really sorry for any inconvenience it may have caused.
