### 2024-09-13 kaipartmann/Peridynamics.jl#167 @ranocha
It looks like it works locally for me. You can close this issue once you have released a new version of Peridynamics.jl including the fix and have updated the JOSS paper branch accordingly.
