### 2024-09-13 #165 @ranocha
https://github.com/kaipartmann/Peridynamics.jl/issues/167 is close to be closed. When that's done, a new release has been made, and the paper branch has been updated, I am fine accepting the paper. Thanks to @kaipartmann and collaborators for a nice contribution to the open-source ecosystem 👍
