### 2024-09-13 530e5e4
Add v0.3.3 to the paper branch
