+++ test/Aqua.jl
@@ -1,0 +1,5 @@
+
+@testitem "Aqua.jl" begin
+    using Aqua
+    Aqua.test_all(Peridynamics; stale_deps=(ignore=[:ThreadPinning],))
+end
