### 2024-09-13 #165 @kaipartmann
I closed the issue and included the current release in the paper branch. Thanks to @ranocha, for your help and invaluable feedback!
