### 2024-09-13 #165 @kaipartmann
@luraess, I've updated the software version. Please use the editorial bot command to update it. It seems only you have the permission to do so. Thank you!
