### 2024-09-16 f1052d6
added tests for 3 untested lines
