+++ test/discretization/test_point_generators.jl
@@ -78,6 +78,30 @@ end
     ]
     @test vol == fill(0.26808257310632905, 1)
 
+    pos, vol = round_sphere(0.2, 0.5)
+    @test pos ≈ [
+        0.0
+        0.0
+        0.0
+    ]
+    @test vol == fill(0.004188790204786391, 1)
+
+    pos, vol = round_sphere(1.1283791670955126, 1.0)
+    @test pos ≈ [
+        0.0
+        0.0
+        0.0
+    ]
+    @test vol == fill(0.7522527780636749, 1)
+
+    pos, vol = round_sphere(0.5641895835477563, 1.0)
+    @test pos ≈ [
+        0.0
+        0.0
+        0.0
+        ]
+    @test vol == fill(0.09403159725795936, 1)
+
     pos, vol = round_sphere(1.6, 0.5; center_position=(1, 2, 3))
     @test pos ≈ [
         1.0 1.0 1.475333726295699 1.2376668631478493 0.7623331368521504 0.5246662737043011 0.7623331368521508 1.2376668631478496 1.492143194265497 1.306846262678719 0.8904878367717978 0.5565943034167349 0.5565943034167349 0.8904878367717977 1.306846262678719 1.0;
@@ -93,14 +117,6 @@ end
         3.0 3.0 3.0
     ]
     @test vol == fill(0.17453292519943295, 3)
-
-    pos, vol = round_sphere(0.8, 0.5; center_position=(1, 2, 3))
-    @test pos ≈ [
-        1.0
-        2.0
-        3.0
-    ]
-    @test vol == fill(0.26808257310632905, 1)
 end
 
 @testitem "round_cylinder" begin
@@ -119,4 +135,11 @@ end
         0.0 0.0 0.0 0.5 0.5 0.5 1.0 1.0 1.0 1.5 1.5 1.5 2.0 2.0 2.0
     ]
     @test vol == fill(0.20943951023931953, 15)
+
+    pos, vol = round_cylinder(1, 2, 1)
+    @test pos ≈ [
+        0 0 0;
+        0 0 0;
+        -1 0 1
+    ]
 end
