+++ test/discretization/test_point_generators.jl
@@ -86,21 +86,21 @@ end
     ]
     @test vol == fill(0.004188790204786391, 1)
 
-    pos, vol = round_sphere(1.1283791670955126, 1.0)
+    pos, vol = round_sphere(2.1, 1.0)
     @test pos ≈ [
-        0.0
-        0.0
-        0.0
+        0.3175426480542942 -0.3175426480542942 -0.3175426480542942 0.3175426480542942;
+        0.3175426480542942 -0.3175426480542942 0.3175426480542942 -0.3175426480542942;
+        0.3175426480542942 0.3175426480542942 -0.3175426480542942 -0.3175426480542942
     ]
-    @test vol == fill(0.7522527780636749, 1)
+    @test vol == fill(1.2122620652039615, 4)
 
-    pos, vol = round_sphere(0.5641895835477563, 1.0)
+    pos, vol = round_sphere(1.8, 1.0)
     @test pos ≈ [
-        0.0
-        0.0
-        0.0
+        -0.4  0.4
+         0.0  0.0
+         0.0  0.0
         ]
-    @test vol == fill(0.09403159725795936, 1)
+    @test vol == fill(1.5268140296446395, 2)
 
     pos, vol = round_sphere(1.6, 0.5; center_position=(1, 2, 3))
     @test pos ≈ [
