### 2024-09-16 efc73df
adjust syntax of existing point generator functions for more consistency
