+++ test/discretization/test_point_generators.jl
@@ -7,7 +7,7 @@
     ]
     @test vol == fill(0.125, 8)
 
-    pos, vol = uniform_box(1, 1, 1, 0.5; center_position=(1, 1, 1))
+    pos, vol = uniform_box(1, 1, 1, 0.5; center=(1, 1, 1))
     @test pos == [
         0.75  1.25  0.75  1.25  0.75  1.25  0.75  1.25
         0.75  0.75  1.25  1.25  0.75  0.75  1.25  1.25
@@ -25,7 +25,7 @@ end
     ]
     @test vol == fill(0.125, 8)
 
-    pos, vol = uniform_sphere(1, 0.5; center_position=(1, 1, 1))
+    pos, vol = uniform_sphere(1, 0.5; center=(1, 1, 1))
     @test pos == [
         0.75  1.25  0.75  1.25  0.75  1.25  0.75  1.25
         0.75  0.75  1.25  1.25  0.75  0.75  1.25  1.25
@@ -43,7 +43,7 @@ end
     ]
     @test vol == fill(0.125, 16)
 
-    pos, vol = uniform_cylinder(1, 2, 0.5; center_position=(1, 2, 1))
+    pos, vol = uniform_cylinder(1, 2, 0.5; center=(1, 2, 1))
     @test pos ≈ [
         0.75  1.25  0.75  1.25  0.75  1.25  0.75  1.25  0.75  1.25  0.75  1.25  0.75  1.25  0.75  1.25
         1.75  1.75  2.25  2.25  1.75  1.75  2.25  2.25  1.75  1.75  2.25  2.25  1.75  1.75  2.25  2.25
@@ -102,7 +102,7 @@ end
         ]
     @test vol == fill(1.5268140296446395, 2)
 
-    pos, vol = round_sphere(1.6, 0.5; center_position=(1, 2, 3))
+    pos, vol = round_sphere(1.6, 0.5; center=(1, 2, 3))
     @test pos ≈ [
         1.0 1.0 1.475333726295699 1.2376668631478493 0.7623331368521504 0.5246662737043011 0.7623331368521508 1.2376668631478496 1.492143194265497 1.306846262678719 0.8904878367717978 0.5565943034167349 0.5565943034167349 0.8904878367717977 1.306846262678719 1.0;
         2.0 2.0 2.0 2.4116510822475945 2.4116510822475945 2.0 1.5883489177524055 1.5883489177524055 2.0 2.384773043159145 2.4798041368797548 2.2135329293091726 1.7864670706908274 1.5201958631202452 1.6152269568408548 2.0;
@@ -110,7 +110,7 @@ end
     ]
     @test vol == fill(0.13404128655316452, 16)
 
-    pos, vol = round_sphere(1, 0.5; center_position=(1, 2, 3))
+    pos, vol = round_sphere(1, 0.5; center=(1, 2, 3))
     @test pos ≈ [
         1.25 0.875 0.875;
         2.0 2.21650635094611 1.7834936490538904;
@@ -128,7 +128,7 @@ end
     ]
     @test vol == fill(0.20943951023931953, 15)
 
-    pos, vol = round_cylinder(1, 2, 0.5; center_position=(1, 2, 1))
+    pos, vol = round_cylinder(1, 2, 0.5; center=(1, 2, 1))
     @test pos ≈ [
         1.25 0.875 0.8749999999999999 1.25 0.875 0.8749999999999999 1.25 0.875 0.8749999999999999 1.25 0.875 0.8749999999999999 1.25 0.875 0.8749999999999999;
[…]
