### 2024-09-16 1f8b5df
applied new point generator syntax to Tutorials
