+++ paper/paper.tex
@@ -52,7 +52,7 @@ The general internal force density for state-based peridynamics is defined as
 \end{equation}
 with the \emph{force vector states} $\vb{t}=\vb{t}(\vb{\Delta X}, t)$ and $\vb{t}'=\vb{t}(-\vb{\Delta X}, t)$.
 In the first original bond-based formulation of peridynamics, the force vector states $\vb{t}$ and  $\vb{t}'$ have the same value and opposite direction.
-This implies intrinsic limitation to only one material parameter and in consequence to restrictions on the Poisson's ratio \cite{Silling2007}.
+In bond-based peridynamics, this implies an intrinsic limitation to only one material parameter and, consequently, restrictions on the Poisson's ratio  \cite{Silling2007}.
 To overcome these restrictions, a state-based peridynamics was established.
 In the ordinary state-based peridynamics, the deformation states of neighboring points also influence the internal force density \cite{Silling2007}.
 This leads to force vector states which are still collinear but not of same value anymore.
