### 2024-10-08 #165 @kaipartmann
Hello @luraess,
Thank you very much for your help and support! We checked the proof and fixed one minor typo. Now, everything looks good except for one minor issue regarding the references. We just accepted the proofs of the article in ref. [14] and now know what DOI the article will have. Should we update the `bib` file, or will this result in an invalid DOI? I think the article of ref. [14] will be published in the next few days.
