### 2024-10-08 #165 @luraess
Hey, can you please try updating the `.bib` file and generate pdf again to see if it builds as expected.
