+++ paper/ref.bib
@@ -124,10 +124,9 @@
 
 @article{Partmann2024PAMM,
   author  = {Kai Partmann and Manuel Dienst and Christian Wieners and Kerstin Weinberg},
-  journal = {submitted to PAMM},
-  number  = {},
+  journal = {PAMM},
   title   = {Peridynamic computations for thin elastic rods},
-  volume  = {},
+  doi     = {10.1002/pamm.202400103},
   year    = {2024}
 }
 
