### 2024-10-09 #165 @kaipartmann
@luraess
The PDF looks good, but as expected, the DOI is invalid as it is only visible to me in the Proofing System of the article.
