+++ src/Peridynamics.jl
@@ -74,7 +74,7 @@ include("auxiliary/mpi.jl")
 
 include("physics/boundary_conditions.jl")
 include("physics/initial_conditions.jl")
-include("physics/material_parameters.jl")
+include("physics/elasticity.jl")
 include("physics/fracture.jl")
 include("physics/short_range_force_contact.jl")
 
@@ -94,7 +94,7 @@ include("core/job.jl")
 include("core/submit.jl")
 include("core/parameter_handler.jl")
 include("core/systems.jl")
-include("core/materials.jl")
+include("core/parameters.jl")
 include("core/storages.jl")
 include("core/time_solvers.jl")
 include("core/halo_exchange.jl")
