+++ src/Peridynamics.jl
@@ -94,7 +94,7 @@ include("core/job.jl")
 include("core/submit.jl")
 include("core/parameter_handler.jl")
 include("core/systems.jl")
-include("core/parameters.jl")
+include("core/material_parameters.jl")
 include("core/storages.jl")
 include("core/time_solvers.jl")
 include("core/halo_exchange.jl")
