+++ src/physics/correspondence.jl
@@ -80,30 +80,39 @@ function Base.show(io::IO, @nospecialize(mat::NOSBMaterial))
     return nothing
 end
 
-struct NOSBPointParameters <: AbstractPointParameters
-    δ::Float64
-    rho::Float64
-    E::Float64
-    nu::Float64
-    G::Float64
-    K::Float64
-    λ::Float64
-    μ::Float64
-    Gc::Float64
-    εc::Float64
+@params NOSBMaterial struct NOSBPointParameters <: AbstractPointParameters
+    @required_params
     bc::Float64
 end
 
-function NOSBPointParameters(::NOSBMaterial, p::Dict{Symbol,Any})
-    δ = get_horizon(p)
-    rho = get_density(p)
-    E, nu, G, K, λ, μ = get_elastic_params(p)
-    Gc, εc = get_frac_params(p, δ, K)
-    bc = 18 * K / (π * δ^4) # bond constant
-    return NOSBPointParameters(δ, rho, E, nu, G, K, λ, μ, Gc, εc, bc)
+function calc_parameter(::NOSBMaterial, ::Val{:bc}, params::NamedTuple)
+    return 18 * params.K / (π * params.δ^4)
 end
 
-@params NOSBMaterial NOSBPointParameters
+# struct NOSBPointParameters <: AbstractPointParameters
+#     δ::Float64
+#     rho::Float64
+#     E::Float64
+#     nu::Float64
+#     G::Float64
+#     K::Float64
+#     λ::Float64
+#     μ::Float64
+#     Gc::Float64
+#     εc::Float64
+#     bc::Float64
+# end
+
+# function NOSBPointParameters(::NOSBMaterial, p::Dict{Symbol,Any})
+#     δ = get_horizon(p)
+#     rho = get_density(p)
+#     E, nu, G, K, λ, μ = get_elastic_params(p)
+#     Gc, εc = get_frac_params(p, δ, K)
+#     bc = 18 * K / (π * δ^4) # bond constant
+#     return NOSBPointParameters(δ, rho, E, nu, G, K, λ, μ, Gc, εc, bc)
+# end
+
+# @params NOSBMaterial NOSBPointParameters
 
 struct NOSBVerletStorage <: AbstractStorage
     position::Matrix{Float64}
