+++ src/core/point_parameters.jl
@@ -10,13 +10,126 @@ function get_point_params(mat::AbstractMaterial, ::Dict{Symbol,Any})
     throw(MethodError(get_point_params, mat))
 end
 
-macro params(material, params)
+function extract_parameters(mat::AbstractMaterial,
+                            custom_params::Vector{Tuple{Symbol,DataType}},
+                            p::Dict{Symbol,Any})
+    required_params = extract_required_parameters(mat, p)
+    params = (; required_params...)
+    for (name, type) in custom_params
+        value = parse_parameter(mat, name, type, p)
+        if isnothing(value)
+            value = calc_parameter(mat, Val(name), required_params)
+        end
+        check_parameter(mat, Val(name), value)
+        params = (; params..., name = value)
+    end
+    return params
+end
+
+function extract_required_parameters(::AbstractMaterial, p::Dict{Symbol,Any})
+    params = (; get_horizon(p)..., get_density(p)..., get_elastic_params(p)...)
+    params = (; params..., get_frac_params(p, params)...)
+    return params
+end
+
+function parse_parameter(::M, keyword::Symbol, ::Type{T}, p::Dict{Symbol,Any}) where {M,T}
+    haskey(p, keyword) || return nothing
+    value::T = convert(T, p[keyword])
+    return value
+end
+
+function calc_parameter(::AbstractMaterial, ::Val{P}, elparams) where {P}
+    msg = "point parameter `$P` is not specified and has no calculation method!\n"
+    return throw(ArgumentError(msg))
+end
+
+function check_parameter(::AbstractMaterial, ::Val{P}, value) where {P}
+    return nothing
+end
+
+function required_point_parameters(::Type{Material}) where {Material}
+    required_params = (:(δ::Float64), # horizon
+                       :(rho::Float64), # density
+                       :(E::Float64), # Young's modulus
+                       :(nu::Float64), # Poisson's ratio
+                       :(G::Float64), # shear modulus
+                       :(K::Float64), # bulk modulus
+                       :(λ::Float64), # Lamé's first parameter
+                       :(μ::Float64), # Lamé's second parameter
+                       :(Gc::Float64), # fracture energy
+                       :(εc::Float64))
+    return required_params
+end
+
+function required_point_parameter_names(::Type{Material}) where {Material}
+    required_params = required_point_parameters(Material)
+    param_names = Tuple(extract_parameter_name(field) for field in required_params)
+    return param_names
+end
+
+function required_point_parameter_types(::Type{Material}) where {Material}
+    required_params = required_point_parameters(Material)
+    param_types = Tuple(extract_parameter_type(field) for field in required_params)
+    return param_types
+end
+
+function extract_parameter_name(expr::Expr)::Symbol
+    if expr.head !== :(::)
+        throw(ArgumentError("invalid expression for parameter extraction!\n"))
+    end
+    return expr.args[1]
+end
+
+function extract_parameter_type(expr::Expr)::DataType
+    if expr.head !== :(::)
+        throw(ArgumentError("invalid expression for parameter extraction!\n"))
+    end
+    return eval(expr.args[2])
+end
+
+function expand_point_parameters(material, struct_expr)
+    # get the struct name and the input fields
+    struct_header = struct_expr.args[2]
+    input_fields = filter(field -> field isa Expr, struct_expr.args[3].args)
+    # construct the fields block
+    required_params = required_point_parameters(material)
+    required_param_names = required_point_parameter_names(material)
+    fields_expr = Vector{Expr}()
+    custom_fields = Vector{Tuple{Symbol,DataType}}()
+    for field in input_fields
+        if field.head === :(::)
+            push!(fields_expr, field)
[…]

[…]
