### 2024-10-09 #165 @luraess
I see. Thanks. If you expect the DOI to become available in the coming days, could we wait and then publish this one. If not, let mw know about an alternative option and we could follow it as well.
