### 2024-10-11 d50c57d
Add interface function to check if struct has required params
