### 2024-10-14 2e6e276
Add some non-working error handling
