+++ src/core/materials.jl
@@ -1,21 +1,21 @@
-function point_param_type(::M) where {M<:AbstractMaterial}
-    return point_param_type(M)
+function point_param_type(mat::AbstractMaterial)
+    return point_param_type(typeof(mat))
 end
 
-function point_param_type(mat::Type{T}) where {T}
-    return throw(MethodError(point_param_type, mat))
+function point_param_type(mat::Type{T}) where {T<:AbstractMaterial}
+    return throw(InterfaceError(mat, point_param_type))
 end
 
-function material_type(::P) where {P<:AbstractPointParameters}
-    return material_type(P)
+function material_type(params::AbstractPointParameters)
+    return material_type(typeof(params))
 end
 
-function material_type(params::Type{T}) where {T}
+function material_type(params::Type{T}) where {T<:AbstractPointParameters}
     return throw(MethodError(material_type, params))
 end
 
 function get_point_params(mat::AbstractMaterial, ::Dict{Symbol,Any})
-    return throw(MethodError(get_point_params, mat))
+    return throw(InterfaceError(mat, get_point_params))
 end
 
 macro params(material, params)
@@ -26,10 +26,10 @@ macro params(material, params)
         Peridynamics.typecheck_params($(esc(params)))
     end
     local _pointparam_type = quote
-        Peridynamics.point_param_type(::Base.Type{<:$(esc(material))}) = $(esc(params))
+        Peridynamics.point_param_type(::Base.Type{M}) where {M<:$(esc(material))} = $(esc(params))
     end
     local _material_type = quote
-        Peridynamics.material_type(::Base.Type{<:$(esc(params))}) = $(esc(material))
+        Peridynamics.material_type(::Base.Type{P}) where {P<:$(esc(params))} = $(esc(material))
     end
     local _get_pointparams = quote
         function Peridynamics.get_point_params(m::$(esc(material)), p::Dict{Symbol,Any})
@@ -65,13 +65,17 @@ function typecheck_params(::Type{Param}) where {Param}
         msg = "$Param is not a valid point parameter type!\n"
         throw(ArgumentError(msg))
     end
-    material = material_type(Param)
-    parameters = fieldnames(Param)
-    for req_param in required_point_parameters(material)
-        if !in(req_param, parameters)
-            msg = "required parameter $req_param not found in $(Param)!\n"
-            error(msg)
-        end
-    end
+    # println("typecheck_param($(Param))")
+    # @show Param
+    # @show material_type(Param)
+    # @show system_type(BBMaterial)
+    # system = system_type(material_type(Param))
+    # parameters = fieldnames(Param)
+    # for req_param in required_point_parameters(system)
+    #     if !in(req_param, parameters)
+    #         msg = "required parameter $req_param not found in $(Param)!\n"
+    #         error(msg)
+    #     end
+    # end
     return nothing
 end
