+++ src/Peridynamics.jl
@@ -92,9 +92,9 @@ include("discretization/body_chunk.jl")
 
 include("core/job.jl")
 include("core/submit.jl")
-include("core/parameter_handler.jl")
 include("core/systems.jl")
-include("core/materials.jl")
+include("core/parameters.jl")
+include("core/parameter_handler.jl")
 include("core/storages.jl")
 include("core/time_solvers.jl")
 include("core/halo_exchange.jl")
