### 2024-10-14 83dac7d
Enhanced error handling regarding interfaces
