+++ src/discretization/interaction_system.jl
@@ -415,3 +415,54 @@ function calc_n_interactions(dh::AbstractMPIBodyDataHandler)
     n_three_nis = MPI.Reduce(length(dh.chunk.system.three_nis), MPI.SUM, mpi_comm())
     return n_one_nis, n_two_nis, n_three_nis
 end
+
+function required_point_parameters(::Type{<:AbstractInteractionSystemMaterial})
+    return (:δ, :rho, elasticity_parameters()..., :C1, :C2, :C3)
+end
+
+function get_required_point_parameters(mat::AbstractInteractionSystemMaterial,
+                                       p::Dict{Symbol,Any})
+    params = (; get_horizon(p)..., get_density(p)..., get_elastic_params(p)...)
+    params = (; params..., get_interaction_parameters(mat, p, params)...)
+    return params
+end
+
+function get_interaction_parameters(mat::AbstractInteractionSystemMaterial,
+                                    p::Dict{Symbol,Any}, params)
+    (; δ, μ, λ) = params
+    if haskey(p, :C1)
+        C1::Float64 = float(p[:C1])
+    else
+        C1 = 0.0
+    end
+
+    if haskey(p, :C2)
+        C2::Float64 = float(p[:C2])
+    else
+        C2 = 0.0
+    end
+
+    if haskey(p, :C3)
+        C3::Float64 = float(p[:C3])
+    else
+        C3 = 0.0
+    end
+
+    if C1 ≈ 0 && C2 ≈ 0 && C3 ≈ 0
+        C1 = 30 / π * μ / δ^4
+        C2 = 0.0
+        C3 = 32 / π^4 * (λ - μ) / δ^12
+    else
+        msg = "interaction parameters for $(typeof(mat)) specified manually!\n"
+        msg *= "Be careful when adjusting these parameters to avoid unexpected outcomes!"
+        @mpiroot @warn msg
+    end
+
+    return (; C1, C2, C3)
+end
+
+function allowed_material_kwargs(::AbstractInteractionSystemMaterial)
+    kwargs = (discretization_kwargs()..., elasticity_kwargs()..., fracture_kwargs()...,
+              :C1, :C2, :C3)
+    return kwargs
+end
