### 2024-10-14 bb8312c
Improved error message
