### 2024-10-14 6275d28
Enhanced error checking for parameter types
