### 2024-10-14 41b7779
Use InterfaceError for consistency
