+++ src/auxiliary/errors.jl
@@ -10,9 +10,9 @@ end
 
 function Base.showerror(io::IO, e::InterfaceError)
     print(io, "interface method not correctly defined!")
-    print(io, "\n  type:      ")
+    print(io, "\n  type:    ")
     printstyled(io, string(e.type); bold=true, color=:red)
-    print(io, "\n  function:  ")
+    print(io, "\n  method:  ")
     printstyled(io, string(e.func); bold=true, color=:red)
     println(io)
     return nothing
