### 2024-10-14 73298bf
Minor cleanup and added tests
