### 2024-10-15 c078557
Updated `bib`-entry for AAM paper
