### 2024-10-15 a442b9f
The `ChunkHandler` is now a part of the system
