### 2024-10-15 1c3d2db
Added new tutorials to `README.md`
