+++ README.md
@@ -82,6 +82,17 @@ in the julia package manager. Please take a look at the [documentation](https://
         <figcaption>Fragmenting cylinder</figcaption><br><img src="https://github.com/user-attachments/assets/58e11123-6143-4e13-8642-7e30c9e6c86d" width="300"/>
    </a></td>
   </tr>
+    <tr>
+    <td align="center">
+      <a href="https://kaipartmann.github.io/Peridynamics.jl/stable/generated/tutorial_wave_in_bar/">
+        <figcaption>Wave propagation</figcaption><br><img src="https://github.com/user-attachments/assets/34165b09-2df9-495d-8e41-acff28d0098c" width="300"/><br>
+      </a>
+    </td>
+    <td align="center">
+      <a href="https://kaipartmann.github.io/Peridynamics.jl/stable/generated/tutorial_wave_interface/">
+        <figcaption>Wave propagation across interface</figcaption><br><img src="https://github.com/user-attachments/assets/c589ab03-e7f5-4b71-b0c2-0f05eb3cdddc" width="300"/>
+   </a></td>
+  </tr>
 </table>
 
 ## Usage
